### 2023-11-21 #6074 @editorialbot
[PRE REVIEW]: Copulas.jl: A fully Distributions.jl-compliant copula package
**Submitting author:** @lrnv (<a href="http://orcid.org/0000-0002-7508-999X">Oskar Laverny</a>)
**Repository:** https://github.com/lrnv/Copulas.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.17
