+++ docs/src/archimedean/generalities.md
@@ -2,39 +2,53 @@
 CurrentModule = Copulas
 ```
 
-# General discussion
+# General Discussion
 
 Archimedean copulas are a large class of copulas that are defined as : ... [ details needed...]
 
-Adding a new `ArchimedeanCopula` is very easy. The `Clayton` implementation is as short as: 
+In our implementation, we separate the generator, in the `Generator` type, from the copula itself. The easyest way to understand this orthogonal design is to look at the Clayton implementation. 
+
+We have the `ClaytonGenerator` that is defined as follows: 
 
 ```julia
-struct ClaytonCopula{d,T} <: ArchimedeanCopula{d}
+struct ClaytonGenerator{T} <: UnivariateGenerator
     θ::T
-    function ClaytonCopula(d,θ)
-        if θ < -1/(d-1)
-            throw(ArgumentError("Theta must be greater than -1/(d-1)"))
-        elseif θ == -1/(d-1)
-            return WCopula(d)
+    function ClaytonGenerator(θ)
+        if θ < -1
+            throw(ArgumentError("Theta must be greater than -1"))
+        elseif θ == -1
+            return WGenerator()
         elseif θ == 0
-            return IndependentCopula(d)
+            return IndependentGenerator()
         elseif θ == Inf
-            return MCopula(d)
+            return MGenerator()
         else
-            return new{d,typeof(θ)}(θ)
+            return new{typeof(θ)}(θ)
         end
     end
 end
-ϕ(  C::ClaytonCopula,      t) = max(1+C.θ*t,zero(t))^(-1/C.θ)
-ϕ⁻¹(C::ClaytonCopula,      t) = (t^(-C.θ)-1)/C.θ
-τ(C::ClaytonCopula) = ifelse(isfinite(C.θ), C.θ/(C.θ+2), 1)
-τ⁻¹(::Type{ClaytonCopula},τ) = ifelse(τ == 1,Inf,2τ/(1-τ))
-williamson_dist(C::ClaytonCopula{d,T}) where {d,T} = C.θ >= 0 ? WilliamsonFromFrailty(Distributions.Gamma(1/C.θ,1),d) : ClaytonWilliamsonDistribution(C.θ,d)
+max_monotony(G::ClaytonGenerator) = G.θ >= 0 ? Inf : Int(floor(1 - 1/G.θ))
+ϕ(  G::ClaytonGenerator, t) = max(1+G.θ*t,zero(t))^(-1/G.θ)
+ϕ⁻¹(G::ClaytonGenerator, t) = (t^(-G.θ)-1)/G.θ
+# ϕ⁽¹⁾(G::ClaytonGenerator, t) =  First derivative of ϕ
+# ϕ⁽ᵏ⁾(G::ClaytonGenerator, k, t) = kth derivative of ϕ
+τ(G::ClaytonGenerator) = ifelse(isfinite(G.θ), G.θ/(G.θ+2), 1)
+τ⁻¹(::Type{T},τ) where T<:ClaytonGenerator = ifelse(τ == 1,Inf,2τ/(1-τ))
+williamson_dist(G::ClaytonGenerator, d) = G.θ >= 0 ? WilliamsonFromFrailty(Distributions.Gamma(1/G.θ,G.θ),d) : ClaytonWilliamsonDistribution(G.θ,d)
+```
+
+
+And then the `ClaytonCopula{d,T}` is simply an alias for `ArchimedeanCopula{d,ClaytonGenerator{T}}`: 
+
+```julia
+const ClaytonCopula{d,T} = ArchimedeanCopula{d,ClaytonGenerator{T}}
+ClaytonCopula(d,θ) = ArchimedeanCopula(d,ClaytonGenerator(θ))
[…]
