+++ docs/src/miscellaneous.md
@@ -6,17 +6,6 @@ CurrentModule = Copulas
 
 A few copulas, while necessary in certain cases and really usefull, are hard to classify. We gether them here for simplicity. 
 
-# Frechet-Hoeffding bounds
-
-The Frechet-Hoeffding bounds are availiable through the following functions: 
-
-```@docs
-MCopula
-```
-```@docs
-WCopula
-```
-
 # Plackett Copula
 
 ```@docs
