+++ docs/src/archimedean/generators.md
@@ -1,0 +1,22 @@
+
+# Archimedean Generators
+
+definition of an archimedean generator... 
+
+why are these stuffs usefull... 
+
+etc... 
+
+
+
+## The Generator API
+
+
+uses williamson-d-transfrom by default. 
+
+
+## The WilliamsonGenerator for Inverse Williamson d-transform
+
+```@docs
+WilliamsonGenerator
+```
