+++ docs/make.jl
@@ -1,9 +1,16 @@
 using Copulas
 using Documenter
+using DocumenterCitations
 
 DocMeta.setdocmeta!(Copulas, :DocTestSetup, :(using Copulas); recursive=true)
 
+bib = CitationBibliography(
+    joinpath(@__DIR__,"src","assets","references.bib"),
+    style=:numeric
+)
+
 makedocs(;
+    plugins=[bib],
     modules=[Copulas],
     authors="Oskar Laverny <oskar.laverny@univ-amu.fr> and contributors",
     repo="https://github.com/lrnv/Copulas.jl/blob/{commit}{path}#{line}",
@@ -11,11 +18,13 @@ makedocs(;
     format=Documenter.HTML(;
         prettyurls=get(ENV, "CI", "false") == "true",
         canonical="https://lrnv.github.io/Copulas.jl",
-        assets=String[],
+        assets=String["assets/citations.css"],
         collapselevel=3,
     ),
     pages=[
-        "Home" => "index.md",
+        "Copulas.jl package" => "index.md",
+        "Getting Started" => "getting_started.md",
+        "Theoretical Background" => "theoretical_background.md",
         "Sklar's Theorem" => "sklar.md",
         "Elliptical Copulas" => [
             "elliptical/generalities.md",
@@ -31,7 +40,9 @@ makedocs(;
             "exemples/fitting_sklar.md",
             "exemples/turing.md",
         ],
-        "Reference" => "reference.md"
+        "Dev Roadmap" => "dev_roadmap.md",
+        "Package Index" => "idx.md",
+        "References" => "references.md",
     ],
 )
 
