+++ docs/src/assets/references.bib
@@ -1,0 +1,638 @@
+@book{cherubini2004,
+  ids = {cherubini2004a},
+  title = {Copula Methods in Finance},
+  author = {Cherubini, Umberto and Luciano, Elisa and Vecchiato, Walter},
+  year = {2004},
+  publisher = {{John Wiley \& Sons}},
+  lccn = {HG106 .C49 2004},
+  keywords = {copula}
+}
+@book{nelsen2006,
+  ids = {nelsen2007,nelsen2007introduction},
+  title = {An Introduction to Copulas},
+  author = {Nelsen, Roger B.},
+  year = {2006},
+  series = {Springer Series in Statistics},
+  edition = {2nd ed},
+  publisher = {{Springer}},
+  address = {{New York}},
+  isbn = {978-0-387-28659-4},
+  langid = {english},
+  lccn = {QA273.6 .N45 2006},
+  keywords = {copula},
+  annotation = {00000}
+}
+@book{joe1997,
+  ids = {joe1997a},
+  title = {Multivariate Models and Multivariate Dependence Concepts},
+  author = {Joe, Harry},
+  year = {1997},
+  publisher = {{CRC press}}
+}
+@book{joe2014,
+  ids = {joe2014a},
+  title = {Dependence Modeling with Copulas},
+  author = {Joe, Harry},
+  year = {2014},
+  publisher = {{CRC press}},
+  keywords = {copula}
+}
+@book{mai2017,
+  title = {Simulating Copulas: Stochastic Models, Sampling Algorithms, and Applications},
+  shorttitle = {Simulating Copulas},
+  author = {Mai, Jan-Frederik and Scherer, Matthias and Czado, Claudia},
+  year = {2017},
+  series = {Series in Quantitative Finance},
+  edition = {2nd edition},
+  number = {vol. 6},
+  publisher = {{World Scientific}},
+  address = {{New Jersey}},
+  isbn = {978-981-314-924-3},
+  langid = {english},
+  lccn = {QA273.6 .M29 2017},
+  keywords = {copula}
+}
+@book{durante2015a,
+  title = {Principles of Copula Theory},
+  author = {Durante, Fabrizio and Sempi, Carlo},
+  year = {2015},
+  publisher = {{Chapman and Hall/CRC}},
+  keywords = {copula}
+}
+@article{durante2017,
+  ids = {durante2017a},
+  title = {The {{Vine Philosopher}}},
+  author = {Durante, Fabrizio and Puccetti, Giovanni and Scherer, Matthias and Vanduffel, Steven},
+  year = {2017},
+  month = dec,
+  journal = {Dependence Modeling},
+  volume = {5},
+  number = {1},
+  pages = {256--267},
+  issn = {2300-2298},
+  langid = {english}
+}
+@book{czado2019,
+  title = {Analyzing {{Dependent Data}} with {{Vine Copulas}}: {{A Practical Guide With R}}},
+  shorttitle = {Analyzing {{Dependent Data}} with {{Vine Copulas}}},
+  author = {Czado, Claudia},
+  year = {2019},
+  series = {Lecture {{Notes}} in {{Statistics}}},
+  volume = {222},
+  publisher = {{Springer International Publishing}},
+  address = {{Cham}},
+  langid = {english},
+  keywords = {copula}
+}
+@article{grosser2021,
+  ids = {grosser2021a},
+  title = {Copulae: {{An}} Overview and Recent Developments},
+  shorttitle = {Copulae},
+  author = {Gr{\"o}{\ss}er, Joshua and Okhrin, Ostap},
+  year = {2021},
+  month = apr,
+  journal = {WIREs Computational Statistics},
+  issn = {1939-5108, 1939-0068},
+  langid = {english}
+}
+@article{sklar1959,
[…]
