+++ docs/src/theoretical_background.md
@@ -1,0 +1,278 @@
+# Copulas and dependence structures
+
+This section gives some general definitions and tools about dependence structures. Along this journey through the mathematical theory of copulas, we take the chance to introduce more specifically several topics that will be exploited in the following chapters.
+The interested reader can take a look at the standard books [joe1997,cherubini2004,nelsen2006,joe2014](@cite) or more recently [mai2017, durante2015a, czado2019,grosser2021](@cite) on the subject of copulas.
+We start by defining a few concepts about dependence structures and copulas.
+
+
+Consider a real valued random vector $\bm X = \left(X_1,...,X_d\right): \Omega \to \mathbb R^d$. The distribution of the random vector $\bm X$ can be characterized by its distribution function  $F$: 
+```math
+\begin{align*}
+  F(\bm x) &= \mathbb P\left(\bm X \le \bm x\right)\\
+  &= \mathbb P\left(\forall i \in \{1,...,d\},\; X_i \le x_i\right).
+\end{align*}
+```
+For a function $F$ to be the distribution function of some random vector, it should be $d$-increasing, right-continuous and left-limited. 
+For $i \in \{1,...,d\}$, the random variables $X_1,...,X_d$, called the marginals of the random vector, also have distribution functions denoted $F_1,...,F_d$ and defined by : 
+```math
+F_i(x_i) = F(+\infty,...,+\infty,x_i,+\infty,...,+\infty).
+```
+
+Note that the range $\mathrm{Ran}(F)$ of a distribution function $F$, univariate or multivariate, is always contained in $[0,1]$. When the random vector or random variable is absolutely continuous with respect to (w.r.t.) the Lebesgue measure restricted to its domain, the range is exactly $[0,1]$. When the distribution is discrete with $n$ atoms, the range is a finite set of $n+1$ values in $[0,1]$.
+
+There is a fundamental functional link between the function $F$ and its marginals $F_1,...,F_d$. This link is expressed by the mean of *copulas. 
+
+> **Definition (Copula) :** A copula, usually denoted $C$, is the distribution function of a random vector with marginals that are all uniform on $[0,1]$, i.e.
+>
+> $C_i(u) = u\mathbb 1_{u \in [0,1]} \text{ for all }i \in 1,...,d.$
+
+At the grounds of the theory of copulas lies Sklar's Theorem [sklar1959](@cite), dating back from 1959.
+
+> **Theorem (Sklar):** For every random vector $\bm X$, there exists a copula $C$ such that 
+>
+> $\forall \bm x\in \mathbb R^d, F(\bm x) = C(F_{1}(x_{1}),...,F_{d}(x_{d})).$
+> The copula $C$ is uniquely determined on $\mathrm{Ran}(F_{1}) \times ... \times \mathrm{Ran}(F_{d})$, where $\mathrm{Ran}(F_i)$ denotes the range of the function $F_i$. In particular, if all marginals are absolutely continuous, $C$ is unique.
+
+
+This result allows to decompose the distribution of $\bm X$ into several components: the marginal distributions on one side, and the copula on the other side, which governs the dependence structure between the marginals. This object is central in our work, and therefore deserves a moment of attention. 
+
+> **Example (Independence):** The function 
+>
+> $\Pi : \bm x \mapsto \prod_{i=1}^d x_i = \bm x^{\bm 1}$ is a copula, corresponding to independent random vectors.
+
+
+Copulas are bounded functions
+
[…]
