+++ docs/Project.toml
@@ -1,3 +1,4 @@
 [deps]
 Copulas = "ae264745-0b69-425e-9d9d-cf662c5eec93"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
