### 2023-11-21 b163883
Remove useless docstring
