+++ Project.toml
@@ -8,7 +8,6 @@ Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 Cubature = "667455a9-e2ce-5579-9412-b964f529a492"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
 MvNormalCDF = "37188c8d-bc69-4638-b057-733e744175ec"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
