### 2023-11-21 81c4b29
remove useless methods
see https://discourse.julialang.org/t/how-to-dispatch-on-a-type-alias/106476/31
