### 2023-11-22 b207f9d
Split string "Copulas.XXXGenerator"
