### 2023-11-22 d59e173
Add better error handling
