### 2023-11-22 4b1746d
Automatic aliasing of archimedeans (#82)
