+++ test/archimedean_tests.jl
@@ -1,6 +1,7 @@
 
-@testitem "Test of τ ∘ τ_inv bijection" begin
+@testitem "Test of τ ∘ τ⁻¹ = Id" begin
     using Random
+    using InteractiveUtils
     using StableRNGs
     rng = StableRNG(123)
 
@@ -14,8 +15,142 @@
     @test check_rnd(AMHCopula,           -0.18, 0.33, 10)
     @test check_rnd(FrankCopula,         -1,    1,    10)
     @test check_rnd(InvGaussianCopula,    0,    1/2,  10)
+
+    # check we did not forget to add a generator: 
+    @test length(InteractiveUtils.subtypes(Copulas.UnivariateGenerator)) == 7
     
 end
+
+@testitem "Test of ρ ∘ ρ⁻¹ = Id" begin
+    using Random
+    using InteractiveUtils
+    using StableRNGs
+    rng = StableRNG(123)
+
+    inv_works(T,rho) = Copulas.ρ(T(2,Copulas.ρ⁻¹(T,rho))) ≈ rho
+    check_rnd(T,min,max,N) = all(inv_works(T,x) for x in min .+ (max-min) .* rand(rng,N))
+
+    # Should be adapted to spearman rho and its inverse when it is possible. 
+    @test_broken check_rnd(ClaytonCopula,       -1,    1,    10)
+    @test_broken check_rnd(GumbelCopula,         0,    1,    10)
+    @test_broken check_rnd(JoeCopula,            0,    1,    10)
+    @test_broken check_rnd(GumbelBarnettCopula, -0.35, 0,    10)
+    @test_broken check_rnd(AMHCopula,           -0.18, 0.33, 10)
+    @test_broken check_rnd(FrankCopula,         -1,    1,    10)
+    @test_broken check_rnd(InvGaussianCopula,    0,    1/2,  10)
+
+    # check we did not forget to add a generator: 
+    @test length(InteractiveUtils.subtypes(Copulas.UnivariateGenerator)) == 7
+end
+
+
+@testitem "Test of τ against empirical value" begin
+    using Random, Distributions
+    using StableRNGs
+    using StatsBase
+    rng = StableRNG(123)
+
+    for C in (
+        AMHCopula(3,0.6)          ,
+        AMHCopula(4,-0.3)         ,
+        ClaytonCopula(2,-0.7)     ,
+        ClaytonCopula(3,-0.1)     ,
+        ClaytonCopula(4,7)        ,
+        FrankCopula(2,-5)         ,
+        FrankCopula(3,12)         ,
+        FrankCopula(4,6)          ,
+        FrankCopula(4,30)         ,
+        FrankCopula(4,37)         ,
+        FrankCopula(4,150)        ,
+        JoeCopula(3,7)            ,
+        GumbelCopula(4,7)         ,
+        GumbelCopula(4,20)        ,
+        GumbelCopula(4,100)       ,
+        GumbelBarnettCopula(3,0.7),
+        InvGaussianCopula(4,0.05) ,
+        InvGaussianCopula(3,8)    ,
+    )
+        d = length(C)
+        K = corkendall(rand(rng,C,1000)')
+        avgτ = (sum(K) .- d) / (d^2-d)
+        @test avgτ ≈ Copulas.τ(C) atol = 0.03
+    end
+end
+
+        
+
+
+
[…]
