+++ test/archimedean_tests.jl
@@ -85,24 +85,24 @@ end
     using Random, Distributions
     using StableRNGs
     rng = StableRNG(123)
-    @test rand(rng, AMHCopula(3,0.6)          ) == [0.6100194232653313, 0.2812460189425596, 0.10333020320732342]
-    @test rand(rng, AMHCopula(4,-0.3)         ) == [0.6794797200689908, 0.2392452398851621, 0.6025427563179961, 0.2844305227337651]
-    @test rand(rng, ClaytonCopula(2,-0.7)     ) == [0.6275520653528824, 0.163616179546588]
-    @test rand(rng, ClaytonCopula(3,-0.1)     ) == [0.3881011762518264, 0.5656986324924199, 0.2940237224994986]
-    @test rand(rng, ClaytonCopula(4,7)        ) == [0.4939026426425556, 0.5327344049793008, 0.508779929573844, 0.458289642358837]
-    @test rand(rng, FrankCopula(2,-5)         ) == [0.7336303867890769, 0.19556711408763564]
-    @test rand(rng, FrankCopula(3,12)         ) == [0.705387567921179, 0.8641668940713177, 0.815244058863104]
-    @test rand(rng, FrankCopula(4,6)          ) == [0.9003228944757687, 0.8260436103707566, 0.9133247903512886, 0.7994891654043613]
-    @test rand(rng, FrankCopula(4,30)         ) == [0.3923476356094088, 0.4382518570496335, 0.39700820876188975, 0.3908555957587312]
-    @test rand(rng, FrankCopula(4,37)         ) == [0.7303681473348445, 0.6898271742903759, 0.7264896311455771, 0.7294073988977854]
-    @test rand(rng, FrankCopula(4,150)        ) == [0.43642655973527883, 0.42188653827683725, 0.41316130988739547, 0.42762193255859465]
-    @test rand(rng, JoeCopula(3,7)            ) == [0.5527859698950577, 0.5839844253671289, 0.6067932545327556]
-    @test rand(rng, GumbelCopula(4,7)         ) == [0.781761681793509, 0.7444203095073173, 0.8307664734410207, 0.7495821288234644]
-    @test rand(rng, GumbelCopula(4,20)        ) == [0.5482947180691102, 0.5460490383319109, 0.5159827606532166, 0.6002166337095293]
-    @test rand(rng, GumbelCopula(4,100)       ) == [0.8199684741926241, 0.81488028576667, 0.8161867411368382, 0.8187239187339411]
-    @test rand(rng, GumbelBarnettCopula(3,0.7)) == [0.33125926946816414, 0.31478129636784674, 0.48594013921379653]
-    @test rand(rng, InvGaussianCopula(4,0.05) ) == [0.4777472572394352, 0.5714072783562308, 0.8312697205666382, 0.5291602862226205]
-    @test rand(rng, InvGaussianCopula(3,8)    ) == [0.5325775326824974, 0.4876131855666078, 0.5207906451760111]
+    @test rand(rng, AMHCopula(3,0.6)          ) ≈ [0.6100194232653313, 0.2812460189425596, 0.10333020320732342]
+    @test rand(rng, AMHCopula(4,-0.3)         ) ≈ [0.6794797200689908, 0.2392452398851621, 0.6025427563179961, 0.2844305227337651]
+    @test rand(rng, ClaytonCopula(2,-0.7)     ) ≈ [0.6275520653528824, 0.163616179546588]
+    @test rand(rng, ClaytonCopula(3,-0.1)     ) ≈ [0.3881011762518264, 0.5656986324924199, 0.2940237224994986]
[…]
