### 2023-11-22 #6074 @lrnv
I skimmed through the list of potential reviewers, and suggestions from the other packages, and arrived at the following list, given in no particular order:

- odow
- AnderGray
- lucaferranti
- Datseris
- jbytecode
- kahaaga
- daviddewhurst
- bnriiitb
- klb2
- vankesteren
- matt-graham
- odunbar
- ManuelStapper
- thomaspinder

Of course if you think of someone else do not hesitate -- those are simply names that meet both the "Julia" side and the "Statistics / Dependence structures / Copulas" side of the submission.

A few
