+++ docs/src/empirical/generalities.md
@@ -1,0 +1,127 @@
+```@meta
+CurrentModule = Copulas
+```
+
+# General Discussion
+
+## Pseudo-observations
+
+
+Through the statistical process leading to the estimation of copulas, one usually observes the data and information on the marginals scale and not on the copula scale. This discrepancy between the observed information and the modeled distribution must be taken into account. A key concept is that of pseudo-observations. 
+
+>**Definition (Pseudo-observations):** If $\bm x \in \mathbb R^{N\times d}$ is an $N$-sample of a $d$-variate real-valued random vector $\bm X$, then the pseudo-observations are the normalized ranks of the marginals of $\bm x$, defined as : 
+>
+> $$\bm u \in \mathbb [0,1]^{N\times d}:\; u_{i,j} = \frac{\mathrm{Rank}(x_{i,j},\,\bm x_{\cdot,j})}{N+1} = \frac{1}{N+1}\sum_{k=1}^N \mathbb 1_{x_{k,j} \le x_{i,j}},$$
+>
+> where $\mathrm{Rank}(y,\bm x)  = \sum\limits_{x_i \in \bm x} \mathbb 1_{x_i \le y}$.
+
+
+In `Copulas.jl`, we provide a function `pseudos` that implement this transformation directly. 
+
+```@docs
+pseudos
+```
+
+## Deheuvel's empirical copula
+
+From these pseudo-observations, an empirical copula is defined as follows:
+
+>**Definition (Deheuvel's empirical copula [deheuvels1979](@cite)):** The empirical distribution function of the normalized ranks,
+>
+>$$\hat{C}_N(\bm u) = \frac{1}{N} \sum_{i=1}^N \mathbb 1_{\bm u_i \le \bm u},$$ is called the empirical copula function.
+
+>**Theorem (Exhaustivity and consistency [deheuvels1979](@cite)):** $\hat{C}_N$ is an exhaustive estimator of $C$, and moreover for any normalizing constants $\{\phi_N, N\in \mathbb N\}$ such that $\lim\limits_{N \to \infty} \phi_N \sqrt{N^{-1}\ln \ln N} = 0$, 
+>
+>$$\lim\limits_{N\to\infty} \phi_N \sup_{\bm u \in [0,1]^d} \lvert\hat{C}_N(\bm u) - C(\bm u) \rvert = 0 \text{ a.s.}$$
+
+$\hat{C}_N$ then converges (weakly) to $C$, the true copula of the random vector $\bm X$, when the number of observations $N$ goes to infinity. 
+
+!!! note "The empirical copula is not a true copula"
+    Despite its name, $\hat{C}_N$ is not a copula since it does not have uniform marginals. Be carrefull. 
+
+In the package, this copula is implemented as the `EmpiricalCopula`: 
+
+```@docs; canonical=false
+EmpiricalCopula
+```
+
+## Beta copula
+
+
+The empirical copula function is not a copula. An easy way to fix this problem is to smooth out the marginals with beta distribution functions: 
+
+
+>**Definition (Beta Copula [segers2017](@cite)):** Denoting $F_{n,r}(x) = \sum_{s=r}^n \binom{n}{s} x^s(1-x)^{n-s}$ the distribution function of a $\mathrm{Beta}(r,n+1-r)$ random variable, the function
+>
+> $$\hat{C}_N^\beta : \bm x \mapsto \frac{1}{N} \sum_{i=1}^N \prod\limits_{j=1}^d F_{n,(N+1)u_{i,j}}(x_j)$$ is a genuine copula, called the Beta copula. 
+
[…]
