+++ docs/src/miscellaneous.md
@@ -2,30 +2,21 @@
 CurrentModule = Copulas
 ```
 
-# Miscellaneous Copulas
+# Other Copulas
 
-A few copulas, while necessary in certain cases and really usefull, are hard to classify. We gether them here for simplicity. 
+A few copulas, while necessary in certain cases and really useful, are hard to classify. We gather them here for simplicity. 
 
-# Plackett Copula
+# `PlackettCopula`
 
 ```@docs
 PlackettCopula
 ```
 
-# Farlie-Gumbel-Morgenstern (FGM) copula
+# `FGMCopula`
+
+Farlie-Gumbel-Morgenstern (FGM) copula
+
 
 ```@docs
 FGMCopula
-```
-
-## Survival Copula
-
-```@docs
-SurvivalCopula
-```
-
-# Empirical Copulas
-
-```@docs
-EmpiricalCopula
 ```

[…]
