+++ docs/src/transformations.md
@@ -1,0 +1,12 @@
+```@meta
+CurrentModule = Copulas
+```
+
+# Transformed Copulas 
+
+
+## `SurvivalCopula`
+
+```@docs
+SurvivalCopula
+```
