### 2023-11-23 22a4bae
add the PowerGenerator to the docs
