### 2023-11-24 a2910fc
Adding a few methods to univariate distributions
