+++ src/LiouvilleCopula.jl
@@ -24,7 +24,7 @@ struct LiouvilleCopula{d,TG} <: Copula{d}
     function LiouvilleCopula(α::Vector{Int},G::Generator)
         d = length(α)
         @assert sum(α) <= max_monotony(G) "The generator you provided is not monotonous enough (the monotony index must be greater than sum(α), and thus this copula does not exists."
-        return new{d, typeof(G)}(G)
+        return new{d, typeof(G)}(Tuple(α), G)
     end
 end
 williamson_dist(C::LiouvilleCopula{d,TG}) where {d,TG} = williamson_dist(C.G,sum(C.α))
@@ -33,23 +33,35 @@ function Distributions._rand!(rng::Distributions.AbstractRNG, C::CT, x::Abstract
     # By default, we use the williamson sampling. 
     r = Distributions.rand(williamson_dist(C))
     for i in 1:length(C)
-        x[i] = Distributions.rand(rng,Gamma(C.α[i],1))
+        x[i] = Distributions.rand(rng,Distributions.Gamma(C.α[i],1))
         x[i] = x[i] * r
         x[i] = Distributions.cdf(williamson_dist(C.G,C.α[i]),x[i])
     end
     return x
 end
-function _cdf(C::CT,u) where {CT<:LiouvilleCopula}
+function _cdf(C::CT,v) where {CT<:LiouvilleCopula}
     d = length(C)
-
-    sx = sum(Distributions.quantile(williamson_dist(C.G,C.α[i]),u[i]) for i in 1:d)
+    if all(v == 1)
+        return one(eltype(v))
+    end
+    if any(v == 0)
+        return zero(eltype(v))
+    end
+    u = 1 .- v # Because survival.
+    Gis = [williamson_dist(C.G,C.α[i]) for i in 1:d]
+    x = [Distributions.quantile(Gis[i],u[i]) for i in 1:d]
+    sx = sum(x)
     r = zero(eltype(u))
-    for i in CartesianIndices(α)
+    for i in CartesianIndices(C.α)
         ii = (ij-1 for ij in Tuple(i))
         sii = sum(ii)
         fii = prod(factorial.(ii))
         # This version is really not efficient as derivatives of the generator ϕ⁽ᵏ⁾(C.G, sii, sx) could be pre-computed since many sii will be the same and sx does never change. 
-        r += (-1)^sii / fii * ϕ⁽ᵏ⁾(C.G, sii, sx) * prod(x .^i)
+        r += (-1)^sii / fii * ϕ⁽ᵏ⁾(C.G, sii, sx) * prod(x .^ii)
     end
+
+    # r is the survival function, not the cdf ! 
+    # but if we use 1-u that might work. 
+
     return r
 end
