### 2023-11-24 b401385
Add functionality to the logarithmic distributons
