### 2023-11-30 30fcb75
new file:   src/MiscellaneousCopulas/RafteryCopula.jl 	deleted:    src/Plotly_functions.jl 	new file:   test/RafteryTest.jl 	deleted:    test/test_plotly.jl
