### 2023-11-30 8e1aa29
modified:   docs/src/miscellaneous.md 	modified:   src/Copulas.jl 	modified:   src/MiscellaneousCopulas/RafteryCopula.jl
