### 2023-11-30 06cd1a4
change affiliation on paper fixes #92
Not finished yet.
