+++ test/margins_uniformity.jl
@@ -1,9 +1,10 @@
-@testitem "Test samples have uniform maginals in [0,1]" begin
+@testitem "Generic tests on every copulas" begin
     using HypothesisTests, Distributions, Random
+    using InteractiveUtils
+    using ForwardDiff
     using StableRNGs
     rng = StableRNG(123)
     cops = (
-        # true represent the fact that cdf(williamson_dist(C),x) is defined or not. 
         AMHCopula(3,0.6),
         AMHCopula(4,-0.3),
         ClaytonCopula(2,-0.7),
@@ -27,17 +28,82 @@
         PlackettCopula(2.0),
         # Others ? Yes probably others too ! 
     )
+
+
+
+    #### Try to ensure that every copula in the package is indeed in this list, to remmember contributors to add their model here: 
+    function _subtypes(type::Type)
+        out = Any[]
+        _subtypes!(out, type)
+    end
+    function _subtypes!(out, type::Type)
+        if !isabstracttype(type)
+            push!(out, type)
+        else
+            foreach(T->_subtypes!(out, T), InteractiveUtils.subtypes(type))
+        end
+        out
+    end
+    for CT in _subtypes(Copulas.Copula)
+        # CT is now a concrete subtype of Copula. 
+        # This will check that 
+        @test any(isa(C,CT) for C in cops)
+    end
+
+    #### methods to numerically derivate the pdf from the cdf : 
+    function _v(u,j,uj)
+        return [(i == j ? uj : u[i]) for i in 1:length(u)]
+    end
+    function _der(j,C,u)
+        if j == 1
+            return ForwardDiff.derivative(u1 -> cdf(C,_v(u,1,u1)), u[1])
+        else
+            return ForwardDiff.derivative(uj -> _der(j-1,C,_v(u,j,uj)),u[j])
+        end
+    end
+    function get_numerical_pdf(C,u)
+        _der(length(C),C,u)
+    end
+
+    
+
     n = 1000
     U = Uniform(0,1)
     for C in cops
+
+        d = length(C)
+        CT = Float64
+        for CC in _subtypes(Copulas.Copula)
+            if isa(C,CC)
+                CT = CC
+                break
+            end
+        end
+        @test CT<:Copulas.Copula # otherwise there is an issue.
+
+        @show C, CT
+        
         nfail = 0
         d = length(C)
         spl = rand(rng,C,n)
-        @assert all(0 <= x <= 1 for x in spl)
-        for i in 1:d
-            @test pvalue(ApproximateOneSampleKSTest(spl[i,:], U),tail=:right) > 0.01 # quite weak but enough at these samples sizes to detect really bad behaviors.
 
-            # also test that cdf is behaving correctly: 
+
+        # Check that the cdf has special values at the bounds: 
+        @test cdf(C,zeros(d)) == 0
+        @test cdf(C,ones(d)) == 1
+
+        # Check that the cdf values are in [0,1]
+        @test all(0 .<= cdf(C,spl) .<= 1)
+
+        # Check that samples are in [0,1]:
+        @test all(0 <= x <= 1 for x in spl)
+
+        # Check uniformity of each marginal : 
+        for i in 1:d
+            # On the samples
+            @test pvalue(ApproximateOneSampleKSTest(spl[i,:], U),tail=:right) > 0.01 # this is weak but enough to catch mistakes. 
+
+            # On the cdf: 
             u = ones(d)
             for val in [0,1,rand(10)...]
                 u[i] = val
@@ -47,6 +113,60 @@
                     @test cdf(C,u) ≈ val
                 end
             end
+            # extra check for zeros: 
+            u = rand(d)
+            u[i] = 0
+            @test iszero(cdf(C,u))
         end
-    end
+
+
[…]
