+++ src/MiscellaneousCopulas/RafteryCopula.jl
@@ -19,9 +19,9 @@ where ``u_{(1)}, \\ldots , u_{(d)}`` denote the order statistics of ``u_1, \\ldo
 
 More details about Multivariate Raftery Copula are found in :
 
-    Saali, T., M. Mesfioui, and A. Shabri, 2023: Multivariate Extension of Raftery Copula. Mathematics, 11, 414, https://doi.org/10.3390/math11020414.    
+    [Raftery2023](@cite) Saali, T., M. Mesfioui, and A. Shabri, 2023: Multivariate Extension of Raftery Copula. Mathematics, 11, 414, https://doi.org/10.3390/math11020414. 
     
-    Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.6.
+    [nelsen2006](@cite) Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.6. 
 
 It has a few special cases:
 - When θ = 0, it is the IndependentCopula.
@@ -41,19 +41,17 @@ struct RafteryCopula{d, P} <: Copula{d}
         end
     end
 end
-function _cdf(R::RafteryCopula, u::Vector{T}) where {T}
-    if length(u) != R.d
-        throw(ArgumentError("Dimension mismatch"))
-    end
-    
+Base.eltype(R::RafteryCopula) = eltype(R.θ)
+
+function _cdf(R::RafteryCopula{d,P}, u::Vector{T}) where {d,P,T}
     # Order the vector u
     u_ordered = sort(u)
     
     term1 = u_ordered[1]
-    term2 = (1 - R.θ) * (1 - R.d) / (1 - R.θ - R.d) * prod(u).^(1/(1 - R.θ))
+    term2 = (1 - R.θ) * (1 - d) / (1 - R.θ - d) * prod(u).^(1/(1 - R.θ))
 
     term3 = 0.0
-    for i in 2:R.d
+    for i in 2:d
         prod_prev = prod(u_ordered[1:i-1])
         term3 += R.θ * (1 - R.θ) / ((1 - R.θ - i) * (2 - R.θ - i)) * prod_prev^(1/(1 - R.θ)) * u_ordered[i]^((2 - R.θ - i) / (1 - R.θ))
     end
@@ -62,17 +60,12 @@ function _cdf(R::RafteryCopula, u::Vector{T}) where {T}
     
     return cdf_value
 end
-
-function Distributions._logpdf(R::RafteryCopula, u::Vector{T}) where {T}
-    if length(u) != R.d
-        throw(ArgumentError("Dimension mismatch"))
-    end
-    
-    # Order the vector u 
+function Distributions._logpdf(R::RafteryCopula{d,P}, u::Vector{T}) where {d,P,T}
+    # Order the vector u
     u_ordered = sort(u)
     
-    term_denominator = (1 - R.θ)^(R.d - 1) * (1 - R.θ - R.d)
-    term_numerator = 1 - R.d - R.θ * u_ordered[R.d]^((1 - R.θ - R.d) / (1 - R.θ))
+    term_denominator = (1 - R.θ)^(d - 1) * (1 - R.θ - d)
+    term_numerator = 1 - d - R.θ * u_ordered[d]^((1 - R.θ - d) / (1 - R.θ))
     term_product = prod(u)^((R.θ) / (1 - R.θ))
     
     logpdf_value = log(term_numerator) - log(term_denominator) + log(term_product)
[…]
