### 2023-11-30 0c6c06d
Removing dimension 1 test
