+++ test/RafteryTest.jl
@@ -3,7 +3,6 @@
         @test isa(RafteryCopula(d,0.0), IndependentCopula)
         @test isa(RafteryCopula(d,1.0), MCopula)
     end    
-    @test_throws ArgumentError RafteryCopula(1,0.5)
     @test_throws ArgumentError RafteryCopula(3,-1.5)
     @test_throws ArgumentError RafteryCopula(2, 2.6)
 end
