### 2023-11-30 c86bfb1
Modify docstring slightly
