### 2023-11-30 1920d01
add tests of CDF and PDF RafteryTest.jl
