### 2023-11-30 ec3d0ee
Add Raftery copula (#91)
Co-authored-by: Oskar Laverny <oskar.laverny@gmail.com>
