### 2023-12-01 2f2ac04
Update miscellaneous.md to add bibliography
