+++ src/MiscellaneousCopulas/SurvivalCopula.jl
@@ -45,7 +45,16 @@ function reverse(u,idx)
     reverse!(v,idx)
     return v
 end
-_cdf(C::SurvivalCopula{d,CT,VI},u) where {d,CT,VI} = _cdf(C.C,reverse(u,C.indices))
+function _cdf(C::SurvivalCopula{d,CT,VI},u) where {d,CT,VI}
+    i = C.indices[end]
+    newC = SurvivalCopula(C.C,C.indices[1:end-1])
+    v = deepcopy(u)
+    v[i] = 1 - v[i]
+    r2 = _cdf(newC,v)
+    v[i] = 1
+    r1 = _cdf(newC,v)
+    return r1 - r2
+end 
 Distributions._logpdf(C::SurvivalCopula{d,CT,VI},u) where {d,CT,VI} = Distributions._logpdf(C.C,reverse(u,C.indices))
 function Distributions._rand!(rng::Distributions.AbstractRNG, C::SurvivalCopula{d,CT,VI}, x::AbstractVector{T}) where {d,CT,VI,T}
     Distributions._rand!(rng,C.C,x)
