### 2023-12-04 e019e06
remove useless methods in Generator
