+++ test/margins_uniformity.jl
@@ -1,10 +1,11 @@
 @testitem "Generic tests on every copulas" begin
-    using HypothesisTests, Distributions, Random
+    using HypothesisTests, Distributions, Random, WilliamsonTransforms
     using InteractiveUtils
     using ForwardDiff
     using StableRNGs
-    rng = StableRNG(123)
+
     cops = (
+        IndependentCopula(3),
         AMHCopula(3,0.6),
         AMHCopula(4,-0.3),
         ClaytonCopula(2,-0.7),
@@ -17,7 +18,7 @@
         JoeCopula(3,7),
         GumbelCopula(4,7),
         GumbelCopula(4,20),
-        GumbelCopula(4,100),
+        # GumbelCopula(4,100),
         GumbelBarnettCopula(3,0.7),
         InvGaussianCopula(4,0.05),
         InvGaussianCopula(3,8),
@@ -25,7 +26,11 @@
         TCopula(4, [1 0.5; 0.5 1]),
         FGMCopula(2,1),
         MCopula(4),
+        WCopula(2),
+        ArchimedeanCopula(2,i𝒲(LogNormal(),2)),
         PlackettCopula(2.0),
+        EmpiricalCopula(randn(2,100),pseudo_values=false),
+        SurvivalCopula(ClaytonCopula(2,-0.7),(1,2)),
         # Others ? Yes probably others too ! 
     )
 
@@ -44,13 +49,16 @@
         end
         out
     end
-    for CT in _subtypes(Copulas.Copula)
-        # CT is now a concrete subtype of Copula. 
-        # This will check that 
+    for CT in _subtypes(Copulas.Copula) # Check that every copula type has been used
         @test any(isa(C,CT) for C in cops)
     end
+    for TG in _subtypes(Copulas.Generator) # Check that every generator has been used 
+        @test any(isa(C.G,TG) for C in cops if typeof(C)<:Copulas.ArchimedeanCopula)
+    end
+
 
     #### methods to numerically derivate the pdf from the cdf : 
+    # Not really efficient as in some cases this return zero while the true pdf is clearly not zero. 
     function _v(u,j,uj)
         return [(i == j ? uj : u[i]) for i in 1:length(u)]
     end
@@ -65,108 +73,130 @@
         _der(length(C),C,u)
     end
 
+    # Filter on archimedeans for fitting tests. 
+    function is_archimedean_with_agenerator(CT)
+        if CT<:ArchimedeanCopula
+            GT = Copulas.generatorof(CT)
+            if !isnothing(GT)
+                if !(GT<:Copulas.ZeroVariateGenerator)
+                    if !(GT<:Copulas.WilliamsonGenerator)
+                        return true
+                    end
+                end
+            end
+        end
+        return false
+    end
     
 
     n = 1000
     U = Uniform(0,1)
     for C in cops
-
-        d = length(C)
-        CT = Float64
-        for CC in _subtypes(Copulas.Copula)
-            if isa(C,CC)
-                CT = CC
-                break
-            end
-        end
-        @test CT<:Copulas.Copula # otherwise there is an issue.
-
-        @show C, CT
+        @show C
         
-        nfail = 0
         d = length(C)
+        CT = typeof(C)
+        rng = StableRNG(123)
         spl = rand(rng,C,n)
 
-
+        if !(CT<:TCopula)
         # Check that the cdf has special values at the bounds: 
-        @test cdf(C,zeros(d)) == 0
-        @test cdf(C,ones(d)) == 1
+            @test cdf(C,zeros(d)) == 0
+            @test cdf(C,ones(d)) == 1
 
-        # Check that the cdf values are in [0,1]
[…]
