### 2023-12-04 2330fe4
fix raftery cdf and relax test pvalue...
