### 2023-12-04 0cf206a
Add automatic tests (#93)
Also fixes a few bugs
