+++ test/margins_uniformity.jl
@@ -1,9 +1,11 @@
-@testitem "Test samples have uniform maginals in [0,1]" begin
-    using HypothesisTests, Distributions, Random
+@testitem "Generic tests on every copulas" begin
+    using HypothesisTests, Distributions, Random, WilliamsonTransforms
+    using InteractiveUtils
+    using ForwardDiff
     using StableRNGs
-    rng = StableRNG(123)
+
     cops = (
-        # true represent the fact that cdf(williamson_dist(C),x) is defined or not. 
+        IndependentCopula(3),
         AMHCopula(3,0.6),
         AMHCopula(4,-0.3),
         ClaytonCopula(2,-0.7),
@@ -16,7 +18,7 @@
         JoeCopula(3,7),
         GumbelCopula(4,7),
         GumbelCopula(4,20),
-        GumbelCopula(4,100),
+        # GumbelCopula(4,100),
         GumbelBarnettCopula(3,0.7),
         InvGaussianCopula(4,0.05),
         InvGaussianCopula(3,8),
@@ -24,31 +26,178 @@
         TCopula(4, [1 0.5; 0.5 1]),
         FGMCopula(2,1),
         MCopula(4),
+        WCopula(2),
+        ArchimedeanCopula(2,i𝒲(LogNormal(),2)),
         PlackettCopula(2.0),
+        EmpiricalCopula(randn(2,100),pseudo_values=false),
+        SurvivalCopula(ClaytonCopula(2,-0.7),(1,2)),
         RafteryCopula(2, 0.2),
-        RafteryCopula(3, 0.5)
+        RafteryCopula(3, 0.5),
         # Others ? Yes probably others too ! 
     )
-    n = 1000
-    U = Uniform(0,1)
-    for C in cops
-        nfail = 0
-        d = length(C)
-        spl = rand(rng,C,n)
-        @assert all(0 <= x <= 1 for x in spl)
-        for i in 1:d
-            @test pvalue(ApproximateOneSampleKSTest(spl[i,:], U),tail=:right) > 0.01 # quite weak but enough at these samples sizes to detect really bad behaviors.
 
-            # also test that cdf is behaving correctly: 
-            u = ones(d)
-            for val in [0,1,rand(10)...]
-                u[i] = val
-                if typeof(C)<:TCopula
-                    @test_broken cdf(C,u) ≈ val
-                else
-                    @test cdf(C,u) ≈ val
+
+
+    #### Try to ensure that every copula in the package is indeed in this list, to remmember contributors to add their model here: 
+    function _subtypes(type::Type)
+        out = Any[]
+        _subtypes!(out, type)
+    end
+    function _subtypes!(out, type::Type)
+        if !isabstracttype(type)
+            push!(out, type)
+        else
+            foreach(T->_subtypes!(out, T), InteractiveUtils.subtypes(type))
+        end
+        out
+    end
+    for CT in _subtypes(Copulas.Copula) # Check that every copula type has been used
+        @test any(isa(C,CT) for C in cops)
+    end
+    for TG in _subtypes(Copulas.Generator) # Check that every generator has been used 
+        @test any(isa(C.G,TG) for C in cops if typeof(C)<:Copulas.ArchimedeanCopula)
+    end
+
+
+    #### methods to numerically derivate the pdf from the cdf : 
+    # Not really efficient as in some cases this return zero while the true pdf is clearly not zero. 
+    function _v(u,j,uj)
+        return [(i == j ? uj : u[i]) for i in 1:length(u)]
+    end
+    function _der(j,C,u)
+        if j == 1
+            return ForwardDiff.derivative(u1 -> cdf(C,_v(u,1,u1)), u[1])
+        else
[…]
