### 2023-12-04 c9046d2
Add major precompile workflow
