+++ Project.toml
@@ -11,6 +11,7 @@ ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
 MvNormalCDF = "37188c8d-bc69-4638-b057-733e744175ec"
+PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
@@ -30,6 +31,7 @@ InteractiveUtils = "1.6"
 LinearAlgebra = "1.6"
 LogExpFunctions = "0.3"
 MvNormalCDF = "0.2, 0.3"
+PrecompileTools = "1"
 QuadGK = "2"
 Random = "1.6"
 Roots = "1, 2"
