### 2023-12-04 f1a8606
Add major precompile workflow
