+++ test/margins_uniformity.jl
@@ -94,7 +94,7 @@
     n = 1000
     U = Uniform(0,1)
     for C in cops
-        
+        @show C
         d = length(C)
         CT = typeof(C)
         rng = StableRNG(123)
