### 2023-12-14 54655fd
fix typo in docstring (#101)
