### 2023-12-14 51eeb7d
fix typo in formula
