### 2023-12-14 8c80205
fix typo in formula (#102)
