### 2023-12-21 facae2e
add few methods
