### 2024-01-19 5169a6d
Add logo to readme too
