### 2024-02-04 #6189 @lucaferranti
Hi @osorensen  and @lrnv :wave: , overall very nice work and I definitely recommend acceptance, here are some comments.

### Paper

Overall the paper is clear and gives a good motivation to the package.

- It would be good to add a reference for `DataGenCopulaBased.jl` and `BivariateCopulas.jl`, especially since both are registered in zenodo.
- At line 40 a pretty strong statement is made
> Copulas.jl is clearly the must fully featured, and brings, as a key feature, the complience with the broader ecosystem.

it is not clear to me from tha paper alone that `Copulas.jl` is the most feature complete of the three and would be good to give some more arguments in the paper. Ideally, a table comparing features of the three would be nice to have.

### Documentation

Overall the API is well documented and gives a lot of background knowledge. A few improvement suggestions

- Currently the documentation feels to have more theory than code, it would be nice to expand the `Examples` section with a few more examples of how to use the code, ideally within some applications of how to solve some "real-world" problems. E.g. applying `Copulas.jl` to some popular datasets and concrete example problems.
- In the [turing integration](https://lrnv.github.io/Copulas.jl/stable/exemples/turing/) example, the code ends with a `plot`, would be good to include the resulting image in the documentation. If possible, some visual representation would also be nice to have for the example of the README
- The contribution guidelines section is very short, but I think this is enough. Since you have the colprac badge, you could add a sentence there saying that more info can be found in the ColPrac guide.

### Typos

I spotted a few typos, you may consider using a tool like [typos-cli](https://crates.io/crates/typos-cli) to check more systematically. This could also be setup on CI (actually the julia repo does it) but might be an overkill.

- line 40 of the paper `must -> most`
- line 41 of the paper `complience -> compliance`
- line 46 of the paper `specify -> specifies`
- line 55 `folloiwng -> following` (I also think `consists of` is the grammatically correct form, instead of `is consisting of` but am not native)
- line 63 `exploit -> exploits`
- Documentation navbar `Exemples -> Examples`
- [here](https://lrnv.github.io/Copulas.jl/stable/dependence_measures/) in the "Unfinished work" box, `wtill -> still`

### Unicode

The paper has some issues rendering some unicode characters

![image](https://github.com/openjournals/joss-reviews/assets/49938764/b33ac5f9-9d46-42ec-8e87-2d77be7b1c05)
