### 2024-02-04 #6189 @lrnv
@lucaferranti Thanks a lot for your time and your review. I definitely agree with all the points you mentioned, 100%.

I will split your review into issues on Copulas.jl to be able to start working on each of them with my co-author @Santymax98. @osorensen it might take a while, but I think it is definitely worth it.
