### 2024-02-04 lrnv/Copulas.jl#114 @lrnv
[Paper] JOSS review checklist
This started as a copy of Luca's and Ander's review but now is our checklist of things to do before publishing the paper.

Link to JOSS review : https://github.com/openjournals/joss-reviews/issues/6189

- [x] #115
- [x] #120
- [x] #121
- [x] #120
- [x] #121
- [x] #116
- [x] #135
- [x] #139
- [x] #140
- [x] #141

### Bugs
- [x] #11
- [x] #50
- [x] #105
- [x] #80
- [x] #98

- [x] #88
- [x] #118
- [x] #119
- [x] #145
- [x] #143
- [x] #142
- [ ] #117

### Features

- [ ] #150
- [ ] #151
- [ ] #152

### TODO at the end of the review

- [x] #129
- [x] #163
