### 2024-02-04 lrnv/Copulas.jl#115 @lrnv
[Paper] Add reference to `DataGenCopulaBased.jl` and `BivariateCopulas.jl`, both have zenodo DOIs.
Link to JOSS review : https://github.com/openjournals/joss-reviews/issues/6189
