### 2024-02-04 lrnv/Copulas.jl#116 @lrnv
[Paper] Add comparison table with BivariateCopulas.jl and DatagenCopulaBased.jl
Link to JOSS review : https://github.com/openjournals/joss-reviews/issues/6189

Lucas wrote:

> At line 40 a pretty strong statement is made
> >Copulas.jl is clearly the must fully featured, and brings, as a key feature, the complience with the broader ecosystem.
>
> It is not clear to me from the paper alone that `Copulas.jl` is the most feature complete of the three and would be good to give some more arguments in the paper.
>
> Ideally, a table comparing features of the three would be nice to have.

I agree that a simple table comparing key functionalities would be usefull. We have to decide what to include in that table.
