### 2024-02-04 lrnv/Copulas.jl#117 @lrnv
[Docs] Find new applications examples for the docs
Link to JOSS review : https://github.com/openjournals/joss-reviews/issues/6189

Luca wrote:

> Currently the documentation feels to have more theory than code, it would be nice to expand the `Examples` section with a few more examples of how to use the code, ideally within some applications of how to solve some "real-world" problems. E.g. applying `Copulas.jl` to some popular datasets and concrete example problems.

Thus we need to find a use-case ona popular dataset or concrete problem to add it to the docs.

@Santymax98 Maybe you have in your work one simple use-case that could be shown ?
