### 2024-02-04 lrnv/Copulas.jl#118 @lrnv
[Docs] Detail a bit more the Turing example (add plot). 
Link to JOSS review : https://github.com/openjournals/joss-reviews/issues/6189

Luca wrote:

> In the turing integration example, the code ends with a `plot`, would be good to include the resulting image in the documentation.
>
> If possible, some visual representation would also be nice to have for the example of the README

So first adding the image in the documentation should be easy, the only think is to tell Documenter.jl to actually compile this code (which takes a while, this is why i removed it the other time).

But then plotting a copula with Plots.jl should also be a possibility, a nice thing to have.
