### 2024-02-04 lrnv/Copulas.jl#119 @lrnv
[Contrib guideline] Add a sentence mentioning the colprac guide.
Link to JOSS review : https://github.com/openjournals/joss-reviews/issues/6189

The contribution guidelines section is very short, but I think this is enough. Since you have the colprac badge, you could add a sentence there saying that more info can be found in the ColPrac guide.
