### 2024-02-04 lrnv/Copulas.jl#121 @lrnv
[Paper] Unicode rendering in code blocks is broken.
Link to JOSS review : https://github.com/openjournals/joss-reviews/issues/6189

![image](https://github.com/openjournals/joss-reviews/assets/49938764/b33ac5f9-9d46-42ec-8e87-2d77be7b1c05)
