### 2024-02-04 9987710
CompatHelper: bump compat for TaylorSeries to 0.16, (keep existing compat) (#111)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
