+++ Project.toml
@@ -38,7 +38,7 @@ Roots = "1, 2"
 SpecialFunctions = "2"
 StableRNGs = "1"
 StatsBase = "0.33, 0.34"
-TaylorSeries = "0.12, 0.13, 0.14, 0.15"
+TaylorSeries = "0.12, 0.13, 0.14, 0.15, 0.16"
 Test = "1.6"
 TestItemRunner = "v0.2"
 WilliamsonTransforms = "0.1"
