+++ docs/src/dependence_measures.md
@@ -1,7 +1,7 @@
-# Mesures of dependency
+# Measures of dependency
 
 
-Although the copula is an object that summarizes completely the dependence structure of any random vector, it is an infinite dimensional object and the interpretation of its properties can be difficult when the dimension gets high. Therefore, the literature has come up with some quantifications of the dependence structure that might be used as univariate summaries, of course imperfect, of certain properties of the copula at hand. 
+Although the copula is an object that summarizes completely the dependence structure of any random vector, it is an infinite dimensional object and the interpretation of its properties can be difficult when the dimension gets high. Therefore, the literature has come up with some quantification of the dependence structure that might be used as univariate summaries, of course imperfect, of certain properties of the copula at hand. 
 
 ## Kendall's Tau 
 
@@ -12,31 +12,31 @@ Although the copula is an object that summarizes completely the dependence struc
 Kendall's tau can be obtained through `τ(C::Copula)`. Its value only depends on the dependence structure and not the marginals. 
 
 !!! warn "Multivariate case"
-    There exists several multivariate extensions of Kendall's tau. The one implemented here is the one we just defined what ever the dimension $d$, be carefull ad the normalisation might differ from other places in the litterature.
+    There exists several multivariate extensions of Kendall's tau. The one implemented here is the one we just defined what ever the dimension $d$, be careful as the normalization might differ from other places in the literature.
 
 
 
 ## Spearman's Rho 
 
-> **Definition (Sperman's ρ):** For a copula $C$ with a density $c$, the Sperman's ρ is defined as: 
+> **Definition (Spearman's ρ):** For a copula $C$ with a density $c$, the Spearman's ρ is defined as: 
 >
 > $$\rho = 12 \int C(\bm u) d\bm u -3.$$
 
-Spearman's Rho  can be obtained through `ρ(C::Copula)`. Its value only depends on the dependence structure and not the marginals. 
+Spearman's Rho can be obtained through `ρ(C::Copula)`. Its value only depends on the dependence structure and not the marginals. 
 
 !!! warn "Multivariate case"
-    There exists several multivariate extensions of Spearman's rho. The one implemented here is the one we just defined what ever the dimension $d$, be carefull ad the normalisation might differ from other places in the litterature.
+    There exists several multivariate extensions of Spearman's rho. The one implemented here is the one we just defined what ever the dimension $d$, be careful as the normalization might differ from other places in the literature.
 
 !!! note "Specific values of tau and rho"
     Kendall's $\tau$ and Spearman's $\rho$ have values between -1 and 1, and are -1 in case of complete anticonomotony and 1 in case of comonotony. Moreover, they are 0 in case of independence. This is 
     why we say that they measure the 'strength' of the dependency.
 
 !!! tip "More-that-bivariate cases"
-    These two dependence measures make more sense in the bivariate case than in other cases, and therefore we sometimes refer to the Kendall's matrix or the Spearman's matrix for the collection of bivariate coefficients associated to a multivariate copula. Many copula estimators are based on these coefficients, see e.g. [genest2011,fredricks2007,derumigny2017](@cite).
+    These two dependence measures make more sense in the bivariate case than in other cases, and therefore we sometimes refer to the Kendall's matrix or the Spearman's matrix for the collection of bivariate coefficients associated to a multivariate copula. Many copula estimators are based on these coefficients, see e.g., [genest2011,fredricks2007,derumigny2017](@cite).
 
 ## Tail dependency
 
-Many people are interested in the tail beahvior of their dependence structures. Tail coefficients summarize this tail behavior.
+Many people are interested in the tail behavior of their dependence structures. Tail coefficients summarize this tail behavior.
 
 >**Definition (Tail dependency):** For a copula $C$, we define (when they exist):
 > ```math
[…]
