### 2024-02-04 lrnv/Copulas.jl#120 @lrnv
See #123 for eventual automatic proof reading.
