### 2024-02-04 372e94a
Fixes a few typos (#124)
* Noted typos in the review

* moar typooos

* Small typos everywhere.
