### 2024-02-04 lrnv/Copulas.jl#121 @lrnv
Asked there : https://github.com/openjournals/joss/issues/963 with the list of problematic characters.
