### 2024-02-04 5b8c156
Add JuliaMono font to the paper
