### 2024-02-06 545a8e4
modified:   joss/paper.md     Possible alternative to comparing the 3 packages
