### 2024-02-06 97d415e
Update paper.md
add dates of package bivaratecopula
