### 2024-02-07 a67393f
Empirical Copula
* Empirical Copula
Fixes #105

* Add a test for #105
