### 2024-02-07 6eeeefd
Rewrite Turing.jl example Fixes #118
