### 2024-02-07 8db1a09
Rewrite Turing.jl example (fixes #118)
* Rewrite Turing.jl example
Fixes #118
