### 2024-02-07 56b03d0
Added wontfix and comment in docstring Fixes #11
