### 2024-02-07 43133de
Convert docstring references to the right format and add a few more Fixes #88
