### 2024-02-07 a78a353
Convert docstring references to the right format and add a few more Fixes #88
* Convert docstring references to the right format and add a few more
Fixes #88
