+++ src/Generator/UnivariateGenerator/JoeGenerator.jl
@@ -18,6 +18,9 @@ The [Joe](https://en.wikipedia.org/wiki/Copula_(probability_theory)#Most_importa
 It has a few special cases: 
 - When θ = 1, it is the IndependentCopula
 - When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)
+
+References:
+* [nelsen2006](@cite) Nelsen, Roger B. An introduction to copulas. Springer, 2006.
 """
 struct JoeGenerator{T} <: UnivariateGenerator
     θ::T
