+++ joss/paper.bib
@@ -88,14 +88,6 @@ doi = {10.18637/jss.v039.i09},
   file = {C:\Users\lrnv\Zotero\storage\VMB6EMI8\Joe - 2014 - Dependence modeling with copulas.pdf}
 }
 
-@book{johnson1987multivariate,
-  title={Multivariate statistical simulation: A guide to selecting and generating continuous multivariate distributions},
-  author={Johnson, Mark E},
-  volume={192},
-  year={1987},
-  publisher={John Wiley \& Sons}
-}
-
 @book{durantePrinciplesCopulaTheory2015,
   title = {Principles of Copula Theory},
   author = {Durante, Fabrizio and Sempi, Carlo},
