### 2024-02-07 b03686d
move turing plot around
