### 2024-02-07 b7baf5e
Add a doc page about other usecases. (#132)
