### 2024-02-07 382388c
Update paper.md
modifying tables, Benchmarktools.jl needs to be cited correctly
