### 2024-02-09 b012cfb
Fix a few issues with kendall/spearmann
