### 2024-02-09 4fe7967
Check for boundaries in generic fallback for cdf.
