### 2024-02-09 4e7262a
Fix generics for `cdf`, `\tau` and `\rho` (#133)
