### 2024-02-09 f26207c
adding bibliography and code
