+++ joss/paper.bib
@@ -223,3 +223,18 @@ doi = {10.18637/jss.v039.i09},
   doi          = {10.5281/zenodo.10412898},
   url          = {https://doi.org/10.5281/zenodo.10412898}
 }
+
+@Article{BenchmarkTools,
+author = {{Chen}, Jiahao and {Revels}, Jarrett},
+title = "{Robust benchmarking in noisy environments}",
+journal = {arXiv e-prints},
+keywords = {Computer Science - Performance, 68N30, B.8.1, D.2.5},
+year = 2016,
+month = "Aug",
+eid = {arXiv:1608.04295},
+archivePrefix ={arXiv},
+eprint = {1608.04295},
+primaryClass = {cs.PF},
+adsurl = {https://ui.adsabs.harvard.edu/abs/2016arXiv160804295C},
+adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
