+++ joss/paper.md
@@ -39,47 +39,22 @@ The Julia package `Copulas.jl` brings most standard copula-related features into
 
 The R package `copula` [@r_copula_citation1; @r_copula_citation2; @r_copula_citation3; @r_copula_citation4] is the gold standard when it comes to sampling, estimating, or simply working around dependence structures. However, in other languages, the available tools are not as developed and/or not as recognized. We bridge the gap in the Julian ecosystem with this Julia-native implementation. Due to the very flexible type system in Julia, our code expressiveness and tidiness will increase its usability and maintainability in the long-run. Type-stability allows sampling in arbitrary precision without requiring more code, and Julia's multiple dispatch yields most of the below-described applications.
 
-There are competing packages in Julia, such as [`BivariateCopulas.jl`](https://github.com/AnderGray/BivariateCopulas.jl) [@BivariateCopulas] which only deals with a few models in bivariate settings but has very nice graphs, or [`DatagenCopulaBased.jl`](https://github.com/iitis/DatagenCopulaBased.jl) [@DatagenCopulaBased_1; @DatagenCopulaBased_2; @DatagenCopulaBased_3; @DatagenCopulaBased_4], which only provides sampling and does not have exactly the same models as `Copulas.jl`. While not fully covering out both of these package's functionality (mostly because the three projects chose different copulas to implement), `Copulas.jl` is clearly the most fully featured, and brings, as a key feature, the compliance with the broader ecosystem.
-# Comparison of other packages with copulas.jl
-According to the package documentation, it is possible to summarize some of the most important functionalities of each package and those most needed by the community. Using the following code we manage to obtain the results for each package.
+There are competing packages in Julia, such as [`BivariateCopulas.jl`](https://github.com/AnderGray/BivariateCopulas.jl) [@BivariateCopulas] which only deals with a few models in bivariate settings but has very nice graphs, or [`DatagenCopulaBased.jl`](https://github.com/iitis/DatagenCopulaBased.jl) [@DatagenCopulaBased_1; @DatagenCopulaBased_2; @DatagenCopulaBased_3; @DatagenCopulaBased_4], which only provides sampling and does not have exactly the same models as `Copulas.jl`. While not fully covering out both of these package's functionality (mostly because the three projects chose different implementation paths), `Copulas.jl` brings, as a key feature, the compliance with the broader ecosystem. The following table provides a feature comparison between the three: 
 
-```julia
-# Function to generate "n" random samples from an archimedean copula of dimension `dim`
-function generate_copula_samples(dim)
-    copula = ClaytonCopula(dim, 0.8)
-    return rand(copula, 10^6)
-end
+| Characteristic                                | `Copulas.jl`       | `DatagenCopulaBased.jl`      | `BivariateCopulas.jl`     |
+|-----------------------------------------------|--------------------|------------------------------|---------------------------|
+| `Distributions.jl`'s API | ✔️ | ❌ | ✔️ |
+| Fitting                  | ✔️ | ✔️ | ❌ |
+| Plotting                 | ❌ | ❌ | ✔️ |
+| Available copulas        |     |     |    |
+| - Classic Bivariate      | ✔️ | ✔️ | ✔️ |
+| - Classic Multivariate   | ✔️ | ✔️ | ❌ |
+| - Archimedeans           | ✔️ (All of them) | ⚠️ Selected ones | ⚠️Selected ones |
+| - Obscure Bivariate      | ✔️ | ❌ | ❌ |
+| - Archimedean Chains     | ❌ | ✔️ | ❌ |
 
-# Efficiency test for generating samples from an archimedean copula
-function test_copula_sampling_efficiency(dim)
-    result = @benchmark generate_copula_samples($dim)
[…]
