### 2024-02-10 7cf9beb
[Paper] add comparison tables and benchmark (#127)
* modified:   joss/paper.md
    Possible alternative to comparing the 3 packages

* Update paper.md

add dates of package bivaratecopula

* Update paper.md

modifying tables, Benchmarktools.jl needs to be cited correctly

* adding bibliography and code

* Update paper.md

Add Acknowledgement

* Clean up and write a last version

---------

Co-authored-by: Oskar Laverny <oskar.laverny@gmail.com>
