### 2024-02-10 #6189 @lrnv
There we are; @lucaferranti we are almost done with your review, you can check out the list of tickets we have here : https://github.com/lrnv/Copulas.jl/issues/114 There are only two big tickets remaining, expected time one to two weeks. @osorensen is that acceptable timing ?
