### 2024-02-10 lrnv/Copulas.jl#135 @lucaferranti
[joss paper] benchmark session
I think there's an extra `e9` in the table

![image](https://github.com/lrnv/Copulas.jl/assets/49938764/56d3d088-7da8-4e1d-a325-c2d9faf727f2)

The first row says "in Secods" but `1.15e9 s` would roughly be 36 years, so probably there's an extra `e9` in that column.

A few small improvement suggestions:

- could you include the code to reproduce the benchmarks somewhere in the repository?

- The table might be clearer with each dimension as a row and each package as a column. It would make it easier to compare libraries against each other. Could either split into two smaller tables, one for times and one for memory usage, or have each package column with two subcolumns. I understand anyway this is markdown and not latex, hence there's a forcing upper bound in what can be achieved. So I am also fine with the current format

- In the memory consumption, it seems that the smallest number is 53 MB, might hence be easier to read that column if the memory usage was expressed in MB
