### 2024-02-10 #6189 @lucaferranti
@lrnv looks very good! I think it's approaching convergence and looking forward to see teh final result after the fixes. I left an issue https://github.com/lrnv/Copulas.jl/issues/135 with a few minor improvement suggestions for the benchmarking table
