### 2024-02-10 #6189 @lucaferranti
oh in the comparison table of libraries there's a small typographic issue

![image](https://github.com/openjournals/joss-reviews/assets/49938764/44dbd713-adf3-4894-aea1-d17880c95067)

but I am not sure if that can be fixed in any way in plain markdown :/
