### 2024-02-10 fac5da4
weird markdown tables behavior
