### 2024-02-10 #6189 @osorensen
> There we are; @lucaferranti we are almost done with your review, you can check out the list of tickets we have here : [lrnv/Copulas.jl#114](https://github.com/lrnv/Copulas.jl/issues/114) There are only two big tickets remaining, expected time one to two weeks. @osorensen is that acceptable timing ?

Absolutely! It's also fine if it takes longer time, but in that case it would be great if you kept us updated on the progress.
