### 2024-02-12 75b06de
Fixing the pdf and clearing out the code. Fixes #98
