### 2024-02-12 ebd58ce
Small fix for boundary conditions
