### 2024-02-12 e1d1cb6
Add downgrade CI See https://www.stochasticlifestyle.com/semantic-versioning-semver-is-flawed-and-downgrade-ci-is-required-to-fix-it/
