### 2024-02-12 6b48904
Modify CI for paper too and check on main
