+++ .github/workflows/Downgrade.yml
@@ -16,14 +16,15 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        version: ['1']
+        version: ['1.6']
     steps:
       - uses: actions/checkout@v4
       - uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.version }}
       - uses: cjdoris/julia-downgrade-compat-action@v1
+        # if: ${{ matrix.version == '1.6' }}
         with:
-          skip: Pkg,TOML
+          skip: InteractiveUtils,LinearAlgebra,Random,Test
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
