### 2024-02-12 6bbfe53
Fix test and clean up Project.toml
