+++ Project.toml
@@ -21,12 +21,12 @@ TaylorSeries = "6aa5eb33-94cf-58f4-a9d0-e4b2c4fc25ea"
 WilliamsonTransforms = "48feb556-9bdd-43a2-8e10-96100ec25e22"
 
 [compat]
-Aqua = "v0.8"
+Aqua = "0.8"
 Combinatorics = "1"
 Cubature = "1.5"
 Distributions = "0.25"
 ForwardDiff = "0.10"
-HypothesisTests = "v0.11"
+HypothesisTests = "0.11"
 InteractiveUtils = "1.6"
 LinearAlgebra = "1.6"
 LogExpFunctions = "0.3"
@@ -40,7 +40,7 @@ StableRNGs = "1"
 StatsBase = "0.33, 0.34"
 TaylorSeries = "0.12, 0.13, 0.14, 0.15, 0.16"
 Test = "1.6"
-TestItemRunner = "v0.2"
+TestItemRunner = "0.2"
 WilliamsonTransforms = "0.1"
 julia = "1.6"
 
