+++ test/margins_uniformity.jl
@@ -37,8 +37,6 @@
         # Others ? Yes probably others too ! 
     )
 
-
-
     #### Try to ensure that every copula in the package is indeed in this list, to remmember contributors to add their model here: 
     function _subtypes(type::Type)
         out = Any[]
