### 2024-02-13 b362df6
Modify CI exclusion rules + new lower bound on Williamsontransform
