+++ Project.toml
@@ -41,7 +41,7 @@ StatsBase = "0.33, 0.34"
 TaylorSeries = "0.12, 0.13, 0.14, 0.15, 0.16"
 Test = "1.6"
 TestItemRunner = "0.2"
-WilliamsonTransforms = "0.1.3"
+WilliamsonTransforms = "0.1.4"
 julia = "1.6"
 
 [extras]
