+++ .github/workflows/draft_paper.yml
@@ -1,7 +1,7 @@
 on:
   push:
      paths: 
-       - 'joss/*'
+       - 'joss/**'
 jobs:
   paper:
     runs-on: ubuntu-latest
