+++ Project.toml
@@ -24,7 +24,7 @@ WilliamsonTransforms = "48feb556-9bdd-43a2-8e10-96100ec25e22"
 Aqua = "0.8"
 Combinatorics = "1"
 Cubature = "1.5"
-Distributions = "0.25"
+Distributions = "0.25.46"
 ForwardDiff = "0.10"
 HypothesisTests = "0.11"
 InteractiveUtils = "1.6"
@@ -35,9 +35,9 @@ PrecompileTools = "1"
 QuadGK = "2"
 Random = "1.6"
 Roots = "1, 2"
-SpecialFunctions = "2"
+SpecialFunctions = "1.2, 2"
 StableRNGs = "1"
-StatsBase = "0.33, 0.34"
+StatsBase = "0.32, 0.33, 0.34"
 TaylorSeries = "0.12, 0.13, 0.14, 0.15, 0.16"
 Test = "1.6"
 TestItemRunner = "0.2"
