+++ Project.toml
@@ -25,7 +25,7 @@ Aqua = "0.8"
 Combinatorics = "1"
 Cubature = "1.5"
 Distributions = "0.25.46"
-ForwardDiff = "0.11"
+ForwardDiff = "0.10"
 HypothesisTests = "0.11"
 InteractiveUtils = "1.6"
 LinearAlgebra = "1.6"
@@ -35,7 +35,7 @@ PrecompileTools = "1"
 QuadGK = "2"
 Random = "1.6"
 Roots = "1, 2"
-SpecialFunctions = "1.2, 2"
+SpecialFunctions = "2"
 StableRNGs = "1"
 StatsBase = "0.32, 0.33, 0.34"
 TaylorSeries = "0.12, 0.13, 0.14, 0.15, 0.16"
