+++ Project.toml
@@ -25,7 +25,7 @@ Aqua = "0.8"
 Combinatorics = "1"
 Cubature = "1.5"
 Distributions = "0.25.46"
-ForwardDiff = "0.10"
+ForwardDiff = "0.10.24"
 HypothesisTests = "0.11"
 InteractiveUtils = "1.6"
 LinearAlgebra = "1.6"
