### 2024-02-13 #6189 @AnderGray
Hi @lrnv and @osorensen! A very useful (and needed) tool in Julia. The documentation is also quite comprehensive and well written, and will serve as a good resource for those reading up or learning about copulas. Acceptance is certainly recommended.

## Paper

- This sentence could be confusing `Copulas being fundamentally distributions of random vectors`. Are Copulas distributions of random vectors? Maybe you mean something like: `Since copulas can combine arbitrary distributions to form correlated random vectors …`

- Line 33 `our code expressiveness and tidiness` -> `our code’s …`

- Line 44 `we did not consider …` -> `we have not considered …`

- In the example 1., there is a bug. I think you want either `x = rand(X,1000)` or `D = SklarDist(C,(X1,X2,X3))`

- With the above fix, I still get an error, related to fitting the binomial distribution
```Julia
ERROR: suffstats is not implemented for (Binomial, Vector{Float64}).
Stacktrace:
  [1] error(s::String)
    @ Base ./error.jl:35
  [2] suffstats(dt::Type{Binomial}, xs::Vector{Float64})
    @ Distributions ~/.julia/packages/Distributions/UaWBm/src/genericfit.jl:5
  [3] fit_mle(dt::Type{Binomial}, x::Vector{Float64})
    @ Distributions ~/.julia/packages/Distributions/UaWBm/src/genericfit.jl:27
  [4] fit(dt::Type{Binomial}, x::Vector{Float64})
    @ Distributions ~/.julia/packages/Distributions/UaWBm/src/genericfit.jl:33
  [5] (::Copulas.var"#12#13"{Tuple{Gamma, Normal, Binomial}, Matrix{Float64}})(i::Int64)
    @ Copulas ./cartesian.jl:0
  [6] iterate
    @ ./generator.jl:47 [inlined]
  [7] collect_to!(dest::Vector{Distribution{Univariate, Continuous}}, itr::Base.Generator{UnitRange{Int64}, Copulas.var"#12#13"{Tuple{Gamma, Normal, Binomial}, Matrix{Float64}}}, offs::Int64, st::Int64)
    @ Base ./array.jl:840
  [8] collect_to!(dest::Vector{Gamma{Float64}}, itr::Base.Generator{UnitRange{Int64}, Copulas.var"#12#13"{Tuple{Gamma, Normal, Binomial}, Matrix{Float64}}}, offs::Int64, st::Int64)
    @ Base ./array.jl:848
  [9] collect_to_with_first!(dest::Vector{Gamma{Float64}}, v1::Gamma{Float64}, itr::Base.Generator{UnitRange{Int64}, Copulas.var"#12#13"{Tuple{Gamma, Normal, Binomial}, Matrix{Float64}}}, st::Int64)
    @ Base ./array.jl:818
 [10] collect(itr::Base.Generator{UnitRange{Int64}, Copulas.var"#12#13"{Tuple{Gamma, Normal, Binomial}, Matrix{Float64}}})
    @ Base ./array.jl:792
 [11] _totuple
    @ ./tuple.jl:401 [inlined]
 [12] Tuple
    @ ./tuple.jl:369 [inlined]
 [13] fit(#unused#::Type{SklarDist{FrankCopula, Tuple{Gamma, Normal, Binomial}}}, x::Matrix{Float64})
    @ Copulas ~/Documents/julia/Copulas_review/Copulas.jl/src/SklarDist.jl:70
 [14] top-level scope
    @ REPL[10]:1
```
- `# Although you'll probbaly get a bad fit !` -> typo in `probably`

- line 60 `has seen` -> `have seen`

- `“# Add the loglikelyhood to the model` -> typo in `loglikelihood`

- The `Generator` example fails, perhaps you forgot to export?
```Julia
ERROR: UndefVarError: `Generator` not defined
Stacktrace:
 [1] top-level scope
   @ REPL[5]:1
```

## Docs
I agree with Luca that the docs could benefit from more code examples. Additionally, here’s a couple of typos I’ve found when reading through the docs:

- There is a typo in the folder name: `exemples` -> `examples`
- `for random ve tors and random variables` -> `vectors` in https://github.com/lrnv/Copulas.jl/blob/6d5f5c5f3ca9724a2feba33612e404a0b0388a7e/docs/src/getting_started.md?plain=1#L102

- In the first example, you say: `Increase the number of observations to get a beter fit (or not?)`. Are you getting a bad fit because you've changed the distribution function Maybe it’s better to have a simpler example where you fit re-fit the `Clayton`? Also `beter` -> `better`

- “SUite” -> “[Suite](https://lrnv.github.io/Copulas.jl/stable/getting_started/#SUite)”

- `Anticonomotony` -> `anticomonotony` in https://github.com/lrnv/Copulas.jl/blob/6d5f5c5f3ca9724a2feba33612e404a0b0388a7e/docs/src/dependence_measures.md?plain=1#L36

- `The transformation, wich is called the inverse williamson transformation` -> `which` and `Williamson` in https://github.com/lrnv/Copulas.jl/blob/6d5f5c5f3ca9724a2feba33612e404a0b0388a7e/src/Generator/WilliamsonGenerator.jl#L14

- Missing a close bracket in equation for InvGaussianGenerator in https://github.com/lrnv/Copulas.jl/blob/32d1ac56383f9da386f05d3992c6fe518e447125/src/Generator/UnivariateGenerator/InvGaussianGenerator.jl#L15

## Features

`Copulas.jl` is very feature complete, including many different types of copulas (elliptical, empirical, Archimedean), and also allows you to create your own copulas using generators. The Williamson transformation is also a very nice idea! Although not required for acceptance of this paper/package, there's a couple suggested "good to have" features

- Conditioning: in `BivariateCopulas.jl` we perform conditioning of copulas and joint distributions using numerical finite differencing. Not sure if there's a better way to do it, but it allows you to sample from arbitrary (without a family) copulas and distributions. I think this is needed if you want to do implement `vines`

- [Rossenblat](https://openturns.github.io/openturns/latest/theory/numerical_methods/rosenblatt_transformation.html) (and inverse) transformation is often used in uncertainty analysis (particularly reliability analysis) to transform random variables into independent standard normal variables.

- Plotting / Visualisation
