### 2024-02-13 #6189 @lrnv
Thanks @AnderGray for your time, your review, and your ideas. I definitely agree with all of your comments and suggestions for `Copulas.jl`. Let met first take a moment to move all your comments into separated issues on the repo, after what you will be able to track progress [at this particular link](https://github.com/lrnv/Copulas.jl/issues/114).
