### 2024-02-13 lrnv/Copulas.jl#151 @lrnv
[New feature] Rosenblatt (and inverse) transformations
AnderGray said in Joss's review:

> Rossenblat (and inverse) transformation is often used in uncertainty analysis (particularly reliability analysis) to transform random variables into independent standard normal variables.

----------------

There is a bit of rosenblatt transformation implemented in `BivariateCopulas.jl`. I know this transformation is very useful, I am just not really sure how we can make it generic enough to work for all copulas (on one side) and specific enough for standard models to obtain fast computations.

It is also used to do quasi-random-number generation with a dependence structure, which is really useful for some use-cases.
