### 2024-02-13 lrnv/Copulas.jl#152 @lrnv
[New feature] Plotting / Vizualization
AnderGray suggested in the Joss review to add functionality for Plotting / Visualisation.

---------

Yes, this is indeed a very good idea. We might borow a few plots from `BivariateCopulas.jl` and also revamp #78, but my goal is to be compatible with Plots.jl and StatsPlots.jl, and do that as a package extension so that loading these things is not necessary to use the package. It will also allow to add a lot of graphs to the documenttion.
