### 2024-02-13 eb5c3bc
Correct paper's example Fixes #139
