### 2024-02-13 9fd6733
[Docs] Proof-reading (#154)
* Move from exemples to examples

* Anticonomotony -> anticomonotony

* Typo

* one more typo

* Fix first example
