+++ docs/src/getting_started.md
@@ -110,12 +110,10 @@ X₃ = LogNormal(0,1)
 C = ClaytonCopula(3,0.7) # A 3-variate Clayton Copula with θ = 0.7
 D = SklarDist(C,(X₁,X₂,X₃)) # The final distribution
 
-# This generates a (3,1000)-sized dataset from the multivariate distribution D
-simu = rand(D,1000)
+simu = rand(D,1000) # Generate a dataset
 
-# While the following estimates the parameters of the model from a dataset: 
-D̂ = fit(SklarDist{FrankCopula,Tuple{Gamma,Normal,LogNormal}}, simu)
-# Increase the number of observations to get a beter fit (or not?)  
+# You may estimate a copula using the `fit` function:
+D̂ = fit(SklarDist{ClaytonCopula,Tuple{Gamma,Normal,LogNormal}}, simu)
 ```
 
 !!! info "About fitting methods"
