### 2024-02-14 lrnv/Copulas.jl#152 @alecloudenback
You might just point folks to https://sefffal.github.io/PairPlots.jl/dev/guide/
