### 2024-02-14 lrnv/Copulas.jl#152 @lrnv
@alecloudenback Yes this is the kind of stuff I want in the package extension, the ideal plot here would be:

- bottom triangle plots bivariate contours on the rank scale [0,1]^2
- upper triangle plots bivariate contours on the z-scale (standard Gaussian marginals)
- (for `plot(::SklarDist)`) The diagonal represents the marginal densities with a `plot(::UnivariateDistribution)`, e.g. the one from StatsPlots.
- (for `plot(::Copula)`) Diagonal elements are TBD, maybe simply variable names ? or something else ?

If PairsPlots.jl can be used here, then perfect.
