### 2024-02-14 8412dcb
Add archimedeans examples
