+++ docs/src/archimedean/generalities.md
@@ -27,7 +27,7 @@ In this package, there is an abstract class [`Generator`](@ref) that contains th
 If you do not find the one you need, you may define it yourself by subtyping `Generator`. The API does not ask for much information, which is really convenient. Only the two following methods are required:
 
 * The `ϕ(G::MyGenerator,t)` function returns the value of the archimedean generator itself. 
-* The `max_monotony(G::MyGenerator)` returns its maximum monotony. 
+* The `max_monotony(G::MyGenerator)` returns its maximum monotony, that is the greater integer $d$ making the generator $d$-monotonous.
 
 Thus, a new generator implementation may simply look like:
 
@@ -41,9 +41,38 @@ max_monotony(G::MyGenerator) = Inf
 !!! tip "Win-Win strategy"
     These two functions are enough to sample the corresponding Archimedean copula (see how in the [Inverse Williamson $d$-transforms](@ref w_trans_section) section of the documentation). However, if you know a bit more about your generator, implementing a few more simple methods can largely fasten the algorithms. You'll find more details on these methods in the [`Generator`](@ref) docstring.
 
+
+For example, Here is a graph of a few Clayton Generators: 
+```@example
+using Copulas: ϕ,ClaytonGenerator,IndependentGenerator
+using Plots
+plot( x -> ϕ(ClaytonGenerator(-0.5),x), xlims=(0,5), label="ClaytonGenerator(-0.5)")
+plot!(x -> ϕ(IndependentGenerator(),x), label="IndependentGenerator()")
+plot!(x -> ϕ(ClaytonGenerator(0.5),x), label="ClaytonGenerator(0.5)")
+plot!(x -> ϕ(ClaytonGenerator(1),x), label="ClaytonGenerator(1)")
+plot!(x -> ϕ(ClaytonGenerator(5),x), label="ClaytonGenerator(5)")
+```
+
+And the corresponding inverse functions: 
+
+```@example
+using Copulas: ϕ⁻¹,ClaytonGenerator,IndependentGenerator
+using Plots
+plot( x -> ϕ⁻¹(ClaytonGenerator(-0.5),x), xlims=(0,1), ylims=(0,5), label="ClaytonGenerator(-0.5)")
+plot!(x -> ϕ⁻¹(IndependentGenerator(),x), label="IndependentGenerator()")
+plot!(x -> ϕ⁻¹(ClaytonGenerator(0.5),x), label="ClaytonGenerator(0.5)")
+plot!(x -> ϕ⁻¹(ClaytonGenerator(1),x), label="ClaytonGenerator(1)")
+plot!(x -> ϕ⁻¹(ClaytonGenerator(5),x), label="ClaytonGenerator(5)")
+```
+
 ```@docs
 Generator
 ```
+
+Note that the rate at which these functions are reaching 0 (and their inverse reaching infinity on the left boundary) can vary a lot from one to the other. Note also that the difference between each of them is easier to grasp on the inverse plot. 
+
+
+
 
 ## Williamson d-transform
 
@@ -59,15 +88,13 @@ In this package, we implemented it through the [`WilliamsonGenerator`](@ref) cla
 
 `WilliamsonGenerator(X::UnivariateRandomVariable, d)`.
 
+
 This function computes the Williamson d-transform of the provided random variable $X$ using the [`WilliamsonTransforms.jl`](https://github.com/lrnv/WilliamsonTransforms.jl) package. See [williamson1955multiply, mcneil2009](@cite) for the literature. 
 
 !!! warn "`max_monotony` of Williamson generators"
     The $d$-transform of a positive random variable is $d$-monotonous but not $k$-monotonous for any $k > d$. Its max monotony is therefore $d$. This has a few implications, one of the biggest one is that the $d$-variate Archimedean copula that corresponds has no density. 
     
[…]
