### 2024-02-14 388812d
add examples to the sklar page
