+++ docs/src/sklar.md
@@ -15,7 +15,7 @@ Recall the following theorem from [sklar1959](@cite):
 The implementation we have of this theorem allows building multivariate distributions by specifying separately their marginals and dependence structures as follows:
 
 
-```julia
+```@example 2
 using Copulas, Distributions, Random
 X₁ = Gamma(2,3)
 X₂ = Pareto()
@@ -23,6 +23,20 @@ X₃ = LogNormal(0,1)
 C = ClaytonCopula(3,0.7) # A 3-variate Clayton Copula with θ = 0.7
 D = SklarDist(C,(X₁,X₂,X₃)) # The final distribution
 ```
+
+Although the output is not formatted, the model is constructed, and can be used in different ways: 
+
+```@example 2
+u = rand(D,10)
+```
+
+```@example 2
+pdf(D, u)
+```
+```@example 2
+cdf(D, u)
+```
+
 
 From this construction, the object `D` is a genuine multivariate random vector following `Distributions.jl`'s API, and can be sampled (`rand()`), can have its probability density function and its distribution function evaluated (respectively `pdf` and `cdf`), etc.
 
