### 2024-02-14 d027ec0
add dependencies to the docs environnement
