### 2024-02-14 af32b26
add example to the EllipticalCopula docstring
