### 2024-02-14 3ecdae0
[Docs] Add examples. (#156)
* Remove pull request conputation of the docs.

* Add archimedeans examples

* Add eliptical examples

* refactor fitting example

* add examples in the getting_started page

* typo

* add examples to the sklar page

* add dependencies to the docs environnement

* add examples to the GaussianCopula docstring

* add example to the TCopula docstring

* typo

* add example to the EllipticalCopula docstring

* typo

* typos
