+++ docs/src/elliptical/generalities.md
@@ -4,7 +4,7 @@ CurrentModule = Copulas
 # [General Discussion](@id elliptical_copulas_header)
 ## Elliptical random vectors
 
-The easiest families of copulas are the one derived from known families of random vectors, and the first presented one is, generally, the Elliptical family. 
+The easiest families of copulas are the one derived from known families of random vectors, and the first presented one are, generally, the Elliptical families (in particular, the Gaussian and Student families are very standard in the litterature). 
 
 > **Definition (Spherical and elliptical random vectors):** A random vector $\bm X$ is said to be spherical if for all orthogonal matrix $\bm A \in O_d(\mathbb R)$, $\bm A\bm X \sim \bm X$. 
 >
@@ -22,7 +22,24 @@ This class contains the (multivariate) Normal and Student distributions, and it 
 ```math
 \phi(\bm t) = \prod_{i=1}^n \phi_i(\bm t) = \prod_{i=1}^n \psi_i(\lVert \bm t \rVert_2^2) = \psi(\lVert \bm t \rVert_2^2),
 ```
-which is still a function of only the norm of $\bm t$. To fix ideas, for Gaussian random vectors, $\psi(t) = e^{-\frac{t^2}{2}}$.
+which is still a function of only the norm of $\bm t$. 
+
+To fix ideas, for Gaussian random vectors, $\psi(t) = e^{-\frac{t^2}{2}}$.
+
+!!! note "Sampling with `Distributions.jl`"
+    Elliptical random vectors in the Gaussian and Student families are available from `Distributions.jl`:
+
+    ```@example 3
+    using Distributions
+    Σ = [1 0.5
+        0.5 1] # variance-covariance matrix.
+    ν = 3 # number of degrees of freedom for the student.
+    N = MvNormal(Σ)
+    ```
+
+    ```@example 3
+    T = MvTDist(ν,Σ)
+    ```
 
 ## Elliptical copulas
 
@@ -41,8 +58,59 @@ Moreover, the form of dependence structures that can be reached inside this clas
 On the other hand, there exist performant estimators of high-dimensional covariance matrices, and a large theory is built on the elliptical assumption of high dimensional random vectors, see e.g., [elidan2013,friedman2010,muller2019](@cite) among others. See also [derumigny2022](@cite) for a recent work on nonparametric estimation of the underlying univariate spherical distribution. 
 
 
-!!! note "Discrepancy with the code"
+!!! note "Note on internal implementation"
     If the exposition we just did on characteristic functions of Elliptical random vectors is fundamental to the definition of elliptical copulas, the package does not use this at all to function, and rather rely on the existence of multivariate and corresponding univariate families of distributions in `Distributions.jl`. 
+
+You can obtain these elliptical copulas by the following code: 
+```julia
+using Copulas
+Σ = [1 0.5
+     0.5 1] # variance-covariance matrix.
+ν = 3 # number of degrees of freedom for the student.
+C_N = GaussianCopula(Σ)
+C_T = TCopula(ν,Σ)
+```
+
+As already stated, the underlying code simply applies Sklar. In all generalities, you may define another elliptical copula by the following structure: 
+
+```julia
+struct MyElliptical{d,T} <: EllipticalCopula{d,T}
+    θ:T
+end
+U(::Type{MyElliptical{d,T}}) where {d,T} # Distribution of the univaraite marginals, Normal() for the Gaussian case. 
+N(::Type{MyElliptical{d,T}}) where {d,T} # Distribution of the mutlivariate random vector, MvNormal(\Sigma) for the Gaussian case. 
+```
+
[…]
