### 2024-02-16 #6189 @lrnv
@osorensen this is still on my checklist but we are working on it. @lucaferranti I think this relates to https://github.com/lrnv/Copulas.jl/issues/117 ?

@AnderGray By the way, we are thinking with my coauthor that the feature requests you made in your review will take a long time to solve. They are clearly relevant, and we will keep them as issues to be solved in the future, but maybe they should not be blocking for the publication of the paper. What do you three @AnderGray @lucaferranti and @osorensen think about that ? I am talking specifically of issues https://github.com/lrnv/Copulas.jl/issues/150 https://github.com/lrnv/Copulas.jl/issues/151 and https://github.com/lrnv/Copulas.jl/issues/152
