### 2024-02-16 lrnv/Copulas.jl#151 @lrnv
@AnderGray By the way, would you agree for us to port plotting interface / rosenblatt transforms / other functionalities from `BivariateCopulas.jl` to `Copulas.jl` ? We will have to think about how to credit you of course, maybe you should author the PRs
