+++ docs/src/archimedean/available_models.md
@@ -2,7 +2,7 @@
 CurrentModule = Copulas
 ```
 
-# [Available Models](@id available_archimedean_models)
+# [Archimedean Generators](@id available_archimedean_models)
 
 ## `WilliamsonGenerator`
 ```@docs; canonical=false
