+++ src/UnivariateDistribution/WilliamsonFromFrailty.jl
@@ -10,14 +10,12 @@ function Distributions.rand(rng::Distributions.AbstractRNG, D::WilliamsonFromFra
     return sy/f
 end
 function Distributions.cdf(D::WilliamsonFromFrailty{TF,d}, x::Real) where {TF,d}
-    # how to compute this cdf ? 
     return 1 - Distributions.expectation(
         e -> Distributions.cdf(D.frailty_dist,e/x),
         Distributions.Erlang(d)
     )
 end
 function Distributions.pdf(D::WilliamsonFromFrailty{TF,d}, x::Real) where {TF,d}
-    # how to compute this cdf ? 
     return 1/x^2 * Distributions.expectation(
         e -> Distributions.pdf(D.frailty_dist,e/x),
         Distributions.Erlang(d)
