+++ test/RafteryTest.jl
@@ -21,8 +21,8 @@ end
 
     @test cdf(RafteryCopula(2, 0.8), [0.2, 0.5]) ≈ 0.199432 atol=1e-5
     @test cdf(RafteryCopula(2, 0.5), [0.3, 0.8]) ≈ 0.2817 atol=1e-5
-    @test_broken cdf(RafteryCopula(3, 0.5), [0.1, 0.2, 0.3]) ≈ 0.08325884 atol=1e-5
-    @test_broken cdf(RafteryCopula(3, 0.1), [0.4, 0.8, 0.2]) ≈ 0.120415 atol=1e-5
+    @test cdf(RafteryCopula(3, 0.5), [0.1, 0.2, 0.3]) ≈ 0.08236007 atol=1e-5
+    @test cdf(RafteryCopula(3, 0.1), [0.4, 0.8, 0.2]) ≈ 0.08581997 atol=1e-5    
 
     @test pdf(RafteryCopula(2, 0.8), [0.2, 0.5]) ≈ 0.114055555 atol=1e-4
     @test pdf(RafteryCopula(2, 0.5), [0.3, 0.8]) ≈ 0.6325 atol=1e-4
@@ -37,4 +37,71 @@ end
     F = RafteryCopula(3,0.5)
     samples = rand(rng,F, n_samples)
     @test size(samples) == (3, n_samples)
+end
+
+@testitem "Check against manual version - CDF" begin
+    # https://github.com/lrnv/Copulas.jl/pull/137
+    using Distributions
+    function prueba_CDF(R::Vector{T}, u::Vector{T}) where T
+        # Order the vector u
+        θ = R[1]
+        # println("param:", θ)
+        d = round(Int, R[2])
+        # println("dimension:", d)
+        u_ordered = sort(u)
+        # println("Sorted vector: ", u_ordered)
+        
+        term1 = u_ordered[1]
+        # println("Term 1: ", term1)
+        
+        term2 = ((1 - θ) * (1 -d)) / (1 - θ - d) * prod(u)^(1/(1 - θ))
+        # println("Term 2: ", term2)
+        
+        term3 = 0.0
+        for i in 2:d # <<<<<<<<--------------- This should be 2:d and not 2:length(R) since length(R) is not the dimension. 
+            prod_prev = prod(u_ordered[1:i-1])
+            term3_part = ((θ * (1 - θ)) / ((1 - θ - i) * (2 - θ - i))) * prod_prev^(1/(1 - θ)) * u_ordered[i]^((2 - θ - i) / (1 - θ))
+            # println("Term 3 (part $i): ", term3_part)
+            term3 += term3_part
+        end
+        
+        # Combine the terms to get the cumulative distribution function
+        cdf_value = term1 + term2 - term3
+        # println("Final CDF value: ", cdf_value)
+        
+        return cdf_value
+    end
+    
+    @test prueba_CDF([0.5,3], [0.1,0.2,0.3]) ≈ cdf(RafteryCopula(3,0.5), [0.1,0.2,0.3])
+    @test prueba_CDF([0.8,2], [0.1,0.2]) ≈ cdf(RafteryCopula(2,0.8), [0.1,0.2])
+    @test prueba_CDF([0.2,2], [0.8,0.2]) ≈ cdf(RafteryCopula(2,0.2), [0.8,0.2])
+end
+
+@testitem "Check against manual version - PDF" begin
+    # https://github.com/lrnv/Copulas.jl/pull/137
+    using Distributions
+    function prueba_PDF(R::Vector{T}, u::Vector{T}) where T
+        # Order the vector u
+        θ = R[1]
+        d = round(Int, R[2])
+        u_ordered = sort(u)
[…]
