### 2024-02-20 6d6f654
Add manually computed values
