### 2024-02-20 d8fed1e
[Bug] RafteryCopula fixes. (#137)
* Fixing the pdf and clearing out the code.
Fixes #98

* Small fix for boundary conditions

* Add final test and fix previous values

* Add manually computed values

---------

Co-authored-by: Santymax98
