### 2024-02-20 ccb4433
Update generalities.md
