+++ docs/src/getting_started.md
@@ -4,11 +4,10 @@ CurrentModule = Copulas
 
 ## Multivariate random vectors
 
-This section gives some general definitions and tools about dependence structures. Along this journey through the mathematical theory of copulas, we link to the rest of the documentation for more specific and detailed arguments on particular points, or simply to link the technical documentation of our implementation. 
+This section gives some general definitions and tools about dependence structures, multivariate random vectors and copulas. Along this journey through the mathematical theory of copulas, we link to the rest of the documentation for more specific and detailed arguments on particular points, or simply to the technical documentation of the actual implementation. 
+The interested theoretical reader can take a look at the standard books on the subject [joe1997,cherubini2004,nelsen2006,joe2014](@cite) or more recently [mai2017, durante2015a, czado2019,grosser2021](@cite). 
 
-The interested theoretical reader can take a look at the standard books on the subject [joe1997,cherubini2004,nelsen2006,joe2014](@cite) or more recently [mai2017, durante2015a, czado2019,grosser2021](@cite). We start here by defining a few concepts about dependence structures and copulas.
-
-
+We start here by defining a few concepts about dependence structures and copulas.
 Consider a real valued random vector $\bm X = \left(X_1,...,X_d\right): \Omega \to \mathbb R^d$. The random variables $X_1,...,X_d$ are called the marginals of the random vector $\bm X$. 
 
 !!! info "Constructing random variables in Julia via `Distributions.jl`"
@@ -23,10 +22,10 @@ Consider a real valued random vector $\bm X = \left(X_1,...,X_d\right): \Omega \
     nothing # hide
     ```
     
-    We refer to [Distributions.jl's documentation](https://github.com/JuliaStats/Distributions.jl) for more details on what you can do with these objects, but here we assume that you are familiar with their API.
+    We refer to [Distributions.jl's documentation](https://github.com/JuliaStats/Distributions.jl) for more details on what you can do with these objects. We assume here that you are familiar with their API.
 
 
-The distribution of the random vector $\bm X$ can be characterized by its distribution function $F$: 
+The probability distribution of the random vector $\bm X$ can be characterized by its *distribution function* $F$: 
 ```math
 \begin{align*}
   F(\bm x) &= \mathbb P\left(\bm X \le \bm x\right)\\
@@ -70,7 +69,7 @@ u = rand(C,10)
 cdf(C,u)
 ```
 
-At the grounds of the theory of copulas lies Sklar's Theorem [sklar1959](@cite), dating back from 1959.
+One of the reasons that makes copulas so useful is discovered by Sklar [sklar1959](@cite) in 1959:
 
 > **Theorem (Sklar):** For every random vector $\bm X$, there exists a copula $C$ such that 
 >
@@ -112,19 +111,19 @@ On the other hand, the [`pseudo()`](@ref Pseudo-observations) function computes 
 
 !!! note "Independent random vectors"
 
-    Distributions.jl proposes the [`product_distribution`](https://juliastats.org/Distributions.jl/stable/multivariate/#Product-distributions) function to create those independent random vectors with given marginals. But you'll see that our approach is much more powerfull. 
+    Distributions.jl proposes the [`product_distribution`](https://juliastats.org/Distributions.jl/stable/multivariate/#Product-distributions) function to create those independent random vectors with given marginals. But you can already see that our approach generalizes to other dependence structres, and is thus much powerfull. 
 
-Copulas are bounded functions
[…]
