+++ docs/src/examples/lossalae.md
@@ -1,0 +1,87 @@
+# Loss-Alae fitting example. 
+
+Loss-Alae is a dataset that is provided in the R `copula` package, which documentation can be found [there](https://search.r-project.org/CRAN/refmans/copula/html/loss.html).
+This dataset corresponds to claims received by an insurer, where the two variables `loss` and `alae` respectively correspond to the amount of the loss and to associated expenses.
+There is a certain dependence structure between the two, and the actual distribution generating this data is of course unknown.
+Our task here is to provide a parametric model that approximates this distribution. 
+
+Let us first import the data : 
+```@example 6
+using Copulas, Distributions, Plots
+data = [
+    10.0 24.0 45.0 51.0 60.0 74.0 75.0 78.0 87.0 100.0 115.0 123.0 133.0 140.0 147.0 147.0 165.0 192.0 200.0 300.0 308.0 311.0 326.0 350.0 350.0 400.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 500.0 520.0 578.0 600.0 600.0 634.0 667.0 700.0 750.0 750.0 750.0 750.0 750.0 750.0 750.0 798.0 800.0 833.0 833.0 875.0 900.0 909.0 916.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1000.0 1075.0 1091.0 1100.0 1122.0 1125.0 1200.0 1200.0 1200.0 1250.0 1250.0 1250.0 1250.0 1250.0 1250.0 1250.0 1250.0 1298.0 1300.0 1300.0 1310.0 1333.0 1375.0 1450.0 1456.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1500.0 1522.0 1545.0 1600.0 1633.0 1658.0 1666.0 1666.0 1667.0 1667.0 1701.0 1750.0 1750.0 1750.0 1750.0 1800.0 1800.0 1800.0 1800.0 1800.0 1892.0 1913.0 1937.0 1947.0 1974.0 1980.0 1993.0 1994.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2000.0 2002.0 2018.0 2022.0 2057.0 2067.0 2070.0 2100.0 2125.0 2180.0 2250.0 2250.0 2250.0 2303.0 2312.0 2325
[…]
