### 2024-02-21 a54c8fb
[CI] Change CI policy (#166)
* remove nightly check for release check

* dont compile paper if nothing changed in its folder
