### 2024-02-21 7c5cc3f
Add kendall function example
