+++ docs/src/examples/lambda_viz.md
@@ -1,0 +1,81 @@
+# Empirical Kendall function and Archimedean's λ function.
+
+The Kendall function is an important function in dependence structure analysis. It is defined for a $d$-variate copula $C$ as 
+
+$$K(t) = \mathbb P \left( C(U_1,...,U_d) \le t \right),$$
+
+where $\bm U = \left(U_1,...,U_n\right)$ are drawn according to $C$.
+From a computational point of view, we often do not access to true observations of the random vector $\m U \sim C$ but rather only observations on the marginal scales. 
+
+Suppose for the sake of the argument that we have a multivariate sample on marignal scales $\left(X_{i,j}\right)_{i \in 1,...,d,\; j \in 1,...,n} with dependence structure $C$. 
+A standard way to approximate $K$ is to rather compute 
+$$Z_j = \frac{1}{n-1} \sum_{k \neq j} \bm 1_{X_{i,j} < X_{i,k} \forall i \in 1,...,d}.$$
+
+Indeed, $K$ can be approximated as the empirical distribution function of $Z_1,...,Z_n$. Here is a sketch implementation of this concept:
+```@example lambda
+struct KendallFunction{T}
+    z::Vector{T}
+    function KendallFunction(x)
+    d,n = size(x)
+    z = zeros(n)
+    for i in 1:n
+        for j in 1:n
+            if j ≠ i
+                z[i] += reduce(&, x[:,j] .< x[:,i])
+            end
+        end
+    end
+    z ./= (n-1)
+    sort!(z)
+    return  new{eltype(z)}(z)
+    end
+end
+function (K::KendallFunction)(t)
+    # Then the K function is simply the empirical cdf of the Z sample:
+    return sum(K.z .≤ t)/length(K.z)
+end
+nothing # hide
+```
+
+Let us try it on a random example: 
+
+```@example lambda
+using Copulas, Distributions, Plots
+X = SklarDist(ClaytonCopula(2,2.7),(Normal(),Pareto()))
+x = rand(X,1000)
+K = KendallFunction(x)
+plot(u -> K(u), xlims = (0,1), title="Empirical Kendall function")
+```
+
+One notable detail on the Kendall function is that is does **not** characterize the copula in all generality. On the other hand, for Archimedean copulas, we have:
+$$K(t) = t - \phi'\{\phi^{-1}(t)\} \phi^{-1}(t).$$
+
+Due to this partical relationship, the Kendall function actually characterizes the generator of the archimedean copula. In fact, this relationship is generally expressed in term of a λ function defined as $$\lambda(t) = t - K(t),$$ which, for archimedean copulas, is obviously equal to $\phi'\{\phi^{-1}(t)\} \phi^{-1}(t)$.
+
+Common λ functions can be easily derived by hand for standard archimedean generators. For any archimedean generator in the package, however, it is even easier to let Julia do the derivation. 
+
+Let's try to compare the empirical λ function from our dataset to a few theoretical ones. For that, we setup parameters of the relevant generators to match the kendall τ of the dataset (because we can). We include for the record the independent and completely monotonous cases.
+
+```@example lambda
+using Copulas: ϕ⁽¹⁾, ϕ⁻¹, τ⁻¹, ClaytonGenerator, GumbelGenerator
+using StatsBase: corkendall
+λ(G,t) = ϕ⁽¹⁾(G,ϕ⁻¹(G,t)) * ϕ⁻¹(G,t)
+plot(u -> u - K(u), xlims = (0,1), label="Empirical λ function")
+κ = corkendall(x')[1,2] # empirical kendall tau
+θ_cl = τ⁻¹(ClaytonGenerator,κ)
+θ_gb = τ⁻¹(GumbelGenerator,κ)
+plot!(u -> λ(ClaytonGenerator(θ_cl),u), label="Clayton")
[…]
