### 2024-02-21 711f2b5
[Docs] Add kendall function example
* Add kendall function example
