### 2024-02-23 b3c666d
[Docs] Add imf1/ifm2 example
* Add imf1/ifm2 example
