### 2024-02-23 lrnv/Copulas.jl#117 @lucaferranti
I think the examples section looks much better now and I'm satisfied.

I cannot close this issue, but I've marked the box in the reivew checklist and notified the editor that the paper is ready from my side
