### 2024-02-23 lrnv/Copulas.jl#117 @lrnv
Reoppend because we have a last example to add before pushing a new version fully, will not take long IMHO.
