### 2024-02-23 9600ade
Remove parentheses in citation
Editorial comment on [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6189)
