### 2024-02-23 01748b6
Small typo
Since Archimedes was a person (although long age), I think it's correct to use upper-case A throughout.
