### 2024-02-23 e807fac
Protect uppercase in BibTeX source
