+++ joss/paper.bib
@@ -192,7 +192,7 @@ doi = {10.18637/jss.v039.i09},
 }
 
 @misc{DatagenCopulaBased_3,
-      title={Selected Methods for non-Gaussian Data Analysis}, 
+      title={Selected Methods for non-{Gaussian} Data Analysis}, 
       author={Krzysztof Domino},
       year={2019},
       eprint={1811.10486},
