+++ joss/paper.bib
@@ -39,7 +39,7 @@ url = {http://www.jstatsoft.org/v34/i09/},
 doi = {10.18637/jss.v034.i09},
 }
 @Article{r_copula_citation4,
-title = {Nested Archimedean Copulas Meet {R}: The {nacopula}
+title = {Nested {Archimedean} Copulas Meet {R}: The {nacopula}
     Package},
 author = {{Marius Hofert} and {Martin M\"achler}},
 journal = {Journal of Statistical Software},
