+++ joss/paper.bib
@@ -183,7 +183,7 @@ doi = {10.18637/jss.v039.i09},
 }
 
 @misc{DatagenCopulaBased_2,
-      title={Introducing higher order correlations to marginals' subset of multivariate data by means of Archimedean copulas}, 
+      title={Introducing higher order correlations to marginals' subset of multivariate data by means of {Archimedean} copulas}, 
       author={Krzysztof Domino and Adam Glos},
       year={2018},
       eprint={1803.07813},
