### 2024-02-23 lrnv/Copulas.jl#176 @osorensen
Something wrong with this reference?
Looks like you have multiple lines in `paper.bib` and that parts of a name is missing?

https://github.com/openjournals/joss-reviews/issues/6189
