### 2024-02-23 lrnv/Copulas.jl#177 @osorensen
Can you recognize this one?
I'm not sufficiently acquainted with copulas to know if the comment "# can you recognize this one ?" should be there or not. Could you please explain what it means? It might well be fine.

https://github.com/openjournals/joss-reviews/issues/6189
