### 2024-02-23 e5f8d79
Remove parentheses in citation (#172)
Editorial comment on [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6189)
