### 2024-02-23 f58446e
Small typo (#174)
* Small typo

Since Archimedes was a person (although long age), I think it's correct to use upper-case A throughout.

* Update paper.md

---------

Co-authored-by: Oskar Laverny <oskar.laverny@gmail.com>
Co-authored-by: Oskar Laverny <oskar.laverny@univ-amu.fr>
