### 2024-02-23 687ef43
Protect uppercase in BibTeX source (#175)
* Protect uppercase in BibTeX source

* Update paper.bib

* Update paper.bib
