+++ joss/paper.bib
@@ -39,7 +39,7 @@ url = {http://www.jstatsoft.org/v34/i09/},
 doi = {10.18637/jss.v034.i09},
 }
 @Article{r_copula_citation4,
-title = {Nested Archimedean Copulas Meet {R}: The {nacopula}
+title = {Nested {Archimedean} Copulas Meet {R}: The {nacopula}
     Package},
 author = {{Marius Hofert} and {Martin M\"achler}},
 journal = {Journal of Statistical Software},
@@ -183,7 +183,7 @@ doi = {10.18637/jss.v039.i09},
 }
 
 @misc{DatagenCopulaBased_2,
-      title={Introducing higher order correlations to marginals' subset of multivariate data by means of Archimedean copulas}, 
+      title={Introducing higher order correlations to marginals' subset of multivariate data by means of {Archimedean} copulas}, 
       author={Krzysztof Domino and Adam Glos},
       year={2018},
       eprint={1803.07813},
@@ -192,7 +192,7 @@ doi = {10.18637/jss.v039.i09},
 }
 
 @misc{DatagenCopulaBased_3,
-      title={Selected Methods for non-Gaussian Data Analysis}, 
+      title={Selected Methods for non-{Gaussian} Data Analysis}, 
       author={Krzysztof Domino},
       year={2019},
       eprint={1811.10486},
