### 2024-02-23 lrnv/Copulas.jl#177 @lrnv
This is the generator of the independent copula. Admitedly, this was a pun : if you end up reading this you should recognize it. But it might not be the right tone, you are right, I'll remove this.
