+++ joss/paper.bib
@@ -211,10 +211,7 @@ doi = {10.18637/jss.v039.i09},
 }
 
 @software{BivariateCopulas,
-  author       = {Ander Gray and
-                  Jasper Behrensdorf and
-                  amrods and
-                  Christian Schilling},
+  author       = {Ander Gray and Jasper Behrensdorf and amrods and Christian Schilling},
   title        = {AnderGray/BivariateCopulas.jl: 0.1.5},
   month        = dec,
   year         = 2023,
