### 2024-02-23 c0497e8
Remove pun in paper (keep in docs) (#178)
Fixes #177
Fixes #176
