### 2024-02-23 lrnv/Copulas.jl#176 @lrnv
Oops should not have been closed before @osorensen validate.
