### 2024-02-24 #6189 @lrnv
So :

- ORCIDS and affilition are OK for both authors.
- Release is 0.1.21 as already told to editorialbot.
- Zenodo record for the release is there : https://zenodo.org/records/10699969 with DOI  https://doi.org/10.5281/zenodo.10699969
- I have checked that title and authors list, including orcid, match from the archive to the JOSS paper.
- I have checked that the licence match (MIT Licence)
