### 2024-02-25 b11ac8a
CompatHelper: bump compat for TaylorSeries to 0.17, (keep existing compat) (#181)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
