### 2024-02-29 c1a0d80
Create CITATION.cff
From Joss https://github.com/openjournals/joss-reviews/issues/6189#issuecomment-1970817069
