### 2024-02-29 8bf346e
[Paper] Change citations guidelines to take the paper into account. (#182)
* Create CITATION.cff

From Joss https://github.com/openjournals/joss-reviews/issues/6189#issuecomment-1970817069

* Update CITATION.bib

* Update README.md

* Update index.md
