+++ docs/src/index.md
@@ -38,11 +38,15 @@ The package is available on Julia's general registry, and can be installed as fo
 Please use the following BibTeX if you want to cite this work: 
 
 ```bibtex
-@software{oskar_laverny_2023_10084669,
-  author       = {Oskar Laverny},
-  title        = {Copulas.jl: A fully `Distributions.jl`-compliant copula package},
-  year         = 2022+,
-  doi          = {10.5281/zenodo.6652672},
-  url          = {https://doi.org/10.5281/zenodo.6652672}
+@article{Laverny2024, 
+    doi = {10.21105/joss.06189}, 
+    url = {https://doi.org/10.21105/joss.06189}, 
+    year = {2024}, 
+    publisher = {The Open Journal}, 
+    volume = {9}, 
+    number = {94}, 
+    pages = {6189}, 
+    author = {Oskar Laverny and Santiago Jimenez}, 
+    title = {Copulas.jl: A fully Distributions.jl-compliant copula package}, journal = {Journal of Open Source Software}
 }
 ```
