### 2024-03-04 lrnv/Copulas.jl#114 @lrnv
As the paper is now published, I'll close this.
