### 2024-03-05 lrnv/Copulas.jl#151 @AnderGray
@lrnv Just seeing this. Yes, very happy to give it a go :)
