### 2024-03-05 lrnv/Copulas.jl#151 @lrnv
@AnderGray Thanks, I will go in this direction then as soon as bandwidth becomes available again.
