### 2024-03-05 lrnv/Copulas.jl#150 @kmarray98
I am writing some functions to implement conditional samplers for copulas where there is a simple analytic formula (e.g FGM copula) as part of another project. Would this be useful to add as a start? I am using the formulae in Appendix 1 of the Trivedi and Zimmerman book.
