### 2024-03-24 lrnv/Copulas.jl#150 @lrnv
@kmarray98 I will reconsider this myself,  do you have some kind of base on which I could build up ? Even a draft would help
