### 2024-03-26 lrnv/Copulas.jl#150 @kmarray98
Hi - sure, sorry I have been quite busy since posting, but I do have something. How is best to transfer it (first time trying to contribute something to a package)?
