### 2024-03-26 lrnv/Copulas.jl#150 @lrnv
Since we finally patched the GlobalSensitivity.jl issues, the urgency is not there anymore. If you want to take more time to pause and ponder before sharing, then it's ok for me too.

What's the best sharing way depends on what you have. The most practical is to open a PR: fork the package, push to your own branch and open a PR. If you are not accustomed to Julia package developpement and GitHub Pull Request functionalities however, it can be a bit hard to grasp the first time but it's a tool worth learning !
